refactor(options): extract helper functions to submenu files (Phase 2)
Move helper functions from mod.rs to their respective submenu files:
- system.rs: discover_system_noteskin_choices, language/translated_titles helpers
- graphics.rs: all display mode, renderer, FPS, resolution, aspect ratio helpers
- sound.rs: volume, sample rate, audio backend, device choice helpers
- select_music.rs: brightness, scroll speed, scorebox, chart info, ITL helpers
- machine.rs: preferred style/mode, font, background style, log level helpers
- gameplay.rs: breakdown style, sync graph mode, sync confidence helpers
- advanced.rs: default fail type helpers, ADVANCED_SONG_PARSING_THREADS_ROW_INDEX

All moved functions use pub(in crate::screens::options) visibility.
yes_no_choice_index/yes_no_from_choice remain in mod.rs (used everywhere).
