+++ src/screens/options/mod.rs
@@ -35,7 +35,11 @@ use crate::engine::present::color;
 use crate::engine::present::font;
 use crate::screens::components::shared::screen_bar::{ScreenBarPosition, ScreenBarTitlePlacement};
 use crate::screens::components::shared::{heart_bg, screen_bar, transitions};
-use null_or_die::{BiasKernel, KernelTarget};
+
+mod submenus;
+#[allow(unused_imports)]
+use submenus::*;
+pub use submenus::update_monitor_specs;
 
 /* ---------------------------- transitions ---------------------------- */
 const TRANSITION_IN_DURATION: f32 = 0.4;
@@ -838,14 +842,6 @@ struct SyncPackConfirmState {
     selection: SyncPackSelection,
     active_choice: u8, // 0 = Yes, 1 = No
 }
-
-#[derive(Clone, Debug)]
-struct SoundDeviceOption {
-    label: String,
-    config_index: Option<u16>,
-    sample_rates_hz: Vec<u32>,
-}
-
 // Local fade timing when swapping between main options list and System Options submenu.
 const SUBMENU_FADE_DURATION: f32 = 0.2;
 
@@ -1072,8 +1068,6 @@ const INPUT_BACKEND_CHOICES: &[Choice] = &[literal_choice("Platform Default")];
 const INPUT_BACKEND_INLINE: bool = true;
 #[cfg(not(target_os = "windows"))]
 const INPUT_BACKEND_INLINE: bool = false;
-const SELECT_MUSIC_SCOREBOX_CYCLE_NUM_CHOICES: usize = 4;
-const SELECT_MUSIC_CHART_INFO_NUM_CHOICES: usize = 2;
 
 const SCORE_IMPORT_DONE_OVERLAY_SECONDS: f32 = 1.5;
 const SCORE_IMPORT_ROW_ENDPOINT_INDEX: usize = 0;
@@ -1081,2540 +1075,6 @@ const SCORE_IMPORT_ROW_PROFILE_INDEX: usize = 1;
 const SCORE_IMPORT_ROW_PACK_INDEX: usize = 2;
 const SCORE_IMPORT_ROW_ONLY_MISSING_INDEX: usize = 3;
 const SYNC_PACK_ROW_PACK_INDEX: usize = 0;
-
-#[cfg(target_os = "linux")]
-const SOUND_LINUX_BACKEND_CHOICES: &[Choice] = &[localized_choice("Common", "Auto")];
-
-fn discover_system_noteskin_choices() -> Vec<String> {
-    let mut names = noteskin_parser::discover_itg_skins("dance");
-    if names.is_empty() {
-        names.push(profile::NoteSkin::DEFAULT_NAME.to_string());
-    }
-    names
-}
-
-fn build_sound_device_options() -> Vec<SoundDeviceOption> {
-    let discovered = if audio::is_initialized() {
-        audio::startup_output_devices()
-    } else {
-        Vec::new()
-    };
-    let default_rates = discovered
-        .iter()
-        .find(|dev| dev.is_default)
-        .map(|dev| dev.sample_rates_hz.clone())
-        .unwrap_or_default();
-    let mut options = Vec::with_capacity(discovered.len() + 1);
-    options.push(SoundDeviceOption {
-        label: tr("Common", "Auto").to_string(),
-        config_index: None,
-        sample_rates_hz: default_rates,
-    });
-    for (idx, dev) in discovered.into_iter().enumerate() {
-        let mut label = dev.name.clone();
-        if dev.is_default {
-            label.push_str(&tr("OptionsSound", "DefaultSuffix"));
-        }
-        options.push(SoundDeviceOption {
-            label,
-            config_index: Some(idx as u16),
-            sample_rates_hz: dev.sample_rates_hz,
-        });
-    }
-    options
-}
-
-#[cfg(target_os = "linux")]
-#[inline(always)]
-fn linux_backend_label(backend: config::LinuxAudioBackend) -> std::sync::Arc<str> {
-    match backend {
-        config::LinuxAudioBackend::Auto => tr("Common", "Auto"),
-        config::LinuxAudioBackend::PipeWire => std::sync::Arc::from("PipeWire"),
-        config::LinuxAudioBackend::PulseAudio => std::sync::Arc::from("PulseAudio"),
-        config::LinuxAudioBackend::Jack => std::sync::Arc::from("JACK"),
-        config::LinuxAudioBackend::Alsa => std::sync::Arc::from("ALSA"),
-    }
-}
-
-#[cfg(target_os = "linux")]
-fn build_linux_backend_choices() -> Vec<String> {
-    audio::available_linux_backends()
-        .into_iter()
-        .map(|backend| linux_backend_label(backend).to_string())
-        .collect()
-}
-
-pub const SYSTEM_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::Game,
-        label: lookup_key("OptionsSystem", "Game"),
-        choices: &[localized_choice("OptionsSystem", "DanceGame")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::Theme,
-        label: lookup_key("OptionsSystem", "Theme"),
-        choices: &[localized_choice("OptionsSystem", "SimplyLoveTheme")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::Language,
-        label: lookup_key("OptionsSystem", "Language"),
-        choices: LANGUAGE_CHOICES,
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::LogLevel,
-        label: lookup_key("OptionsSystem", "LogLevel"),
-        choices: &[
-            localized_choice("OptionsSystem", "LogLevelError"),
-            localized_choice("OptionsSystem", "LogLevelWarn"),
-            localized_choice("OptionsSystem", "LogLevelInfo"),
-            localized_choice("OptionsSystem", "LogLevelDebug"),
-            localized_choice("OptionsSystem", "LogLevelTrace"),
-        ],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::LogFile,
-        label: lookup_key("OptionsSystem", "LogFile"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::DefaultNoteSkin,
-        label: lookup_key("OptionsSystem", "DefaultNoteSkin"),
-        choices: &[literal_choice(profile::NoteSkin::DEFAULT_NAME)],
-        inline: false,
-    },
-];
-
-pub const SYSTEM_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::SysGame,
-        name: lookup_key("OptionsSystem", "Game"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSystemHelp",
-            "GameHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SysTheme,
-        name: lookup_key("OptionsSystem", "Theme"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSystemHelp",
-            "ThemeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SysLanguage,
-        name: lookup_key("OptionsSystem", "Language"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSystemHelp",
-            "LanguageHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SysLogLevel,
-        name: lookup_key("OptionsSystem", "LogLevel"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSystemHelp",
-            "LogLevelHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SysLogFile,
-        name: lookup_key("OptionsSystem", "LogFile"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSystemHelp",
-            "LogFileHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SysDefaultNoteSkin,
-        name: lookup_key("OptionsSystem", "DefaultNoteSkin"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSystemHelp",
-            "DefaultNoteSkinHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-#[cfg(all(target_os = "windows", not(target_pointer_width = "32")))]
-const VIDEO_RENDERER_OPTIONS: &[(BackendType, &str)] = &[
-    (BackendType::OpenGL, "OpenGL"),
-    (BackendType::Vulkan, "Vulkan"),
-    (BackendType::DirectX, "DirectX"),
-    (BackendType::OpenGLWgpu, "OpenGL (wgpu)"),
-    (BackendType::VulkanWgpu, "Vulkan (wgpu)"),
-    (BackendType::Software, "Software"),
-];
-#[cfg(all(target_os = "windows", target_pointer_width = "32"))]
-const VIDEO_RENDERER_OPTIONS: &[(BackendType, &str)] = &[
-    (BackendType::OpenGL, "OpenGL"),
-    (BackendType::DirectX, "DirectX"),
-    (BackendType::OpenGLWgpu, "OpenGL (wgpu)"),
-    (BackendType::Software, "Software"),
-];
-#[cfg(all(target_os = "macos", not(target_pointer_width = "32")))]
-const VIDEO_RENDERER_OPTIONS: &[(BackendType, &str)] = &[
-    (BackendType::OpenGL, "OpenGL"),
-    (BackendType::Vulkan, "Vulkan"),
-    (BackendType::Metal, "Metal (wgpu)"),
-    (BackendType::OpenGLWgpu, "OpenGL (wgpu)"),
-    (BackendType::VulkanWgpu, "Vulkan (wgpu)"),
-    (BackendType::Software, "Software"),
-];
-#[cfg(all(
-    not(any(target_os = "windows", target_os = "macos")),
-    not(target_pointer_width = "32")
-))]
-const VIDEO_RENDERER_OPTIONS: &[(BackendType, &str)] = &[
-    (BackendType::OpenGL, "OpenGL"),
-    (BackendType::Vulkan, "Vulkan"),
-    (BackendType::OpenGLWgpu, "OpenGL (wgpu)"),
-    (BackendType::VulkanWgpu, "Vulkan (wgpu)"),
-    (BackendType::Software, "Software"),
-];
-#[cfg(all(not(target_os = "windows"), target_pointer_width = "32"))]
-const VIDEO_RENDERER_OPTIONS: &[(BackendType, &str)] = &[
-    (BackendType::OpenGL, "OpenGL"),
-    (BackendType::OpenGLWgpu, "OpenGL (wgpu)"),
-    (BackendType::Software, "Software"),
-];
-
-#[cfg(all(target_os = "windows", not(target_pointer_width = "32")))]
-const VIDEO_RENDERER_LABELS: &[Choice] = &[
-    localized_choice("OptionsGraphics", "RendererOpenGL"),
-    localized_choice("OptionsGraphics", "RendererVulkan"),
-    localized_choice("OptionsGraphics", "RendererDirectX"),
-    localized_choice("OptionsGraphics", "RendererOpenGLWgpu"),
-    localized_choice("OptionsGraphics", "RendererVulkanWgpu"),
-    localized_choice("OptionsGraphics", "RendererSoftware"),
-];
-#[cfg(all(target_os = "windows", target_pointer_width = "32"))]
-const VIDEO_RENDERER_LABELS: &[Choice] = &[
-    localized_choice("OptionsGraphics", "RendererOpenGL"),
-    localized_choice("OptionsGraphics", "RendererDirectX"),
-    localized_choice("OptionsGraphics", "RendererOpenGLWgpu"),
-    localized_choice("OptionsGraphics", "RendererSoftware"),
-];
-#[cfg(all(target_os = "macos", not(target_pointer_width = "32")))]
-const VIDEO_RENDERER_LABELS: &[Choice] = &[
-    localized_choice("OptionsGraphics", "RendererOpenGL"),
-    localized_choice("OptionsGraphics", "RendererVulkan"),
-    localized_choice("OptionsGraphics", "RendererMetal"),
-    localized_choice("OptionsGraphics", "RendererOpenGLWgpu"),
-    localized_choice("OptionsGraphics", "RendererVulkanWgpu"),
-    localized_choice("OptionsGraphics", "RendererSoftware"),
-];
-#[cfg(all(
-    not(any(target_os = "windows", target_os = "macos")),
-    not(target_pointer_width = "32")
-))]
-const VIDEO_RENDERER_LABELS: &[Choice] = &[
-    localized_choice("OptionsGraphics", "RendererOpenGL"),
-    localized_choice("OptionsGraphics", "RendererVulkan"),
-    localized_choice("OptionsGraphics", "RendererOpenGLWgpu"),
-    localized_choice("OptionsGraphics", "RendererVulkanWgpu"),
-    localized_choice("OptionsGraphics", "RendererSoftware"),
-];
-#[cfg(all(not(target_os = "windows"), target_pointer_width = "32"))]
-const VIDEO_RENDERER_LABELS: &[Choice] = &[
-    localized_choice("OptionsGraphics", "RendererOpenGL"),
-    localized_choice("OptionsGraphics", "RendererOpenGLWgpu"),
-    localized_choice("OptionsGraphics", "RendererSoftware"),
-];
-
-const DISPLAY_ASPECT_RATIO_CHOICES: &[Choice] = &[
-    literal_choice("16:9"),
-    literal_choice("16:10"),
-    literal_choice("4:3"),
-    literal_choice("1:1"),
-];
-
-const VIDEO_RENDERER_ROW_INDEX: usize = 0;
-const SOFTWARE_THREADS_ROW_INDEX: usize = 1;
-const DISPLAY_MODE_ROW_INDEX: usize = 2;
-const DISPLAY_ASPECT_RATIO_ROW_INDEX: usize = 3;
-const DISPLAY_RESOLUTION_ROW_INDEX: usize = 4;
-const REFRESH_RATE_ROW_INDEX: usize = 5;
-const FULLSCREEN_TYPE_ROW_INDEX: usize = 6;
-const VSYNC_ROW_INDEX: usize = 7;
-const PRESENT_MODE_ROW_INDEX: usize = 8;
-const MAX_FPS_ENABLED_ROW_INDEX: usize = 9;
-const MAX_FPS_VALUE_ROW_INDEX: usize = 10;
-const SELECT_MUSIC_SHOW_BANNERS_ROW_INDEX: usize = 0;
-const SELECT_MUSIC_SHOW_VIDEO_BANNERS_ROW_INDEX: usize = 1;
-const SELECT_MUSIC_SHOW_BREAKDOWN_ROW_INDEX: usize = 2;
-const SELECT_MUSIC_BREAKDOWN_STYLE_ROW_INDEX: usize = 3;
-const SELECT_MUSIC_MUSIC_PREVIEWS_ROW_INDEX: usize = 15;
-const SELECT_MUSIC_CHART_INFO_ROW_INDEX: usize = 14;
-const SELECT_MUSIC_PREVIEW_LOOP_ROW_INDEX: usize = 17;
-const SELECT_MUSIC_SHOW_SCOREBOX_ROW_INDEX: usize = 19;
-const SELECT_MUSIC_SCOREBOX_PLACEMENT_ROW_INDEX: usize = 20;
-const SELECT_MUSIC_SCOREBOX_CYCLE_ROW_INDEX: usize = 21;
-const MACHINE_SELECT_STYLE_ROW_INDEX: usize = 2;
-const MACHINE_PREFERRED_STYLE_ROW_INDEX: usize = 3;
-const MACHINE_SELECT_PLAY_MODE_ROW_INDEX: usize = 4;
-const MACHINE_PREFERRED_MODE_ROW_INDEX: usize = 5;
-const ADVANCED_SONG_PARSING_THREADS_ROW_INDEX: usize = 3;
-
-const MAX_FPS_MIN: u16 = 5;
-const MAX_FPS_MAX: u16 = 1000;
-const MAX_FPS_STEP: u16 = 5;
-const MAX_FPS_DEFAULT: u16 = 60;
-const MUSIC_WHEEL_SCROLL_SPEED_VALUES: [u8; 7] = [5, 10, 15, 25, 30, 45, 100];
-
-const DEFAULT_RESOLUTION_CHOICES: &[(u32, u32)] = &[
-    (1920, 1080),
-    (1600, 900),
-    (1280, 720),
-    (1024, 768),
-    (800, 600),
-];
-
-fn build_display_mode_choices(monitor_specs: &[MonitorSpec]) -> Vec<String> {
-    if monitor_specs.is_empty() {
-        return vec![
-            tr("OptionsGraphics", "Screen1Fallback").to_string(),
-            tr("OptionsGraphics", "Windowed").to_string(),
-        ];
-    }
-    let mut out = Vec::with_capacity(monitor_specs.len() + 1);
-    for spec in monitor_specs {
-        out.push(spec.name.clone());
-    }
-    out.push(tr("OptionsGraphics", "Windowed").to_string());
-    out
-}
-
-pub const GRAPHICS_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::VideoRenderer,
-        label: lookup_key("OptionsGraphics", "VideoRenderer"),
-        choices: VIDEO_RENDERER_LABELS,
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::SoftwareRendererThreads,
-        label: lookup_key("OptionsGraphics", "SoftwareRendererThreads"),
-        choices: &[localized_choice("Common", "Auto")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::DisplayMode,
-        label: lookup_key("OptionsGraphics", "DisplayMode"),
-        choices: &[
-            localized_choice("OptionsGraphics", "Windowed"),
-            localized_choice("OptionsGraphics", "Fullscreen"),
-            localized_choice("OptionsGraphics", "Borderless"),
-        ], // Replaced dynamically
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::DisplayAspectRatio,
-        label: lookup_key("OptionsGraphics", "DisplayAspectRatio"),
-        choices: DISPLAY_ASPECT_RATIO_CHOICES,
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::DisplayResolution,
-        label: lookup_key("OptionsGraphics", "DisplayResolution"),
-        choices: &[
-            literal_choice("1920x1080"),
-            literal_choice("1600x900"),
-            literal_choice("1280x720"),
-            literal_choice("1024x768"),
-            literal_choice("800x600"),
-        ], // Replaced dynamically
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::RefreshRate,
-        label: lookup_key("OptionsGraphics", "RefreshRate"),
-        choices: &[
-            localized_choice("Common", "Default"),
-            literal_choice("60 Hz"),
-            literal_choice("75 Hz"),
-            literal_choice("120 Hz"),
-            literal_choice("144 Hz"),
-            literal_choice("165 Hz"),
-            literal_choice("240 Hz"),
-            literal_choice("360 Hz"),
-        ], // Replaced dynamically
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::FullscreenType,
-        label: lookup_key("OptionsGraphics", "FullscreenType"),
-        choices: &[
-            localized_choice("OptionsGraphics", "FullscreenTypeExclusive"),
-            localized_choice("OptionsGraphics", "FullscreenTypeBorderless"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::VSync,
-        label: lookup_key("OptionsGraphics", "VSync"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::PresentMode,
-        label: lookup_key("OptionsGraphics", "PresentMode"),
-        choices: &[literal_choice("Mailbox"), literal_choice("Immediate")],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MaxFps,
-        label: lookup_key("OptionsGraphics", "MaxFps"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MaxFpsValue,
-        label: lookup_key("OptionsGraphics", "MaxFpsValue"),
-        choices: &[localized_choice("Common", "Off")], // Replaced dynamically
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::ShowStats,
-        label: lookup_key("OptionsGraphics", "ShowStats"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("OptionsGraphics", "ShowStatsFPS"),
-            localized_choice("OptionsGraphics", "ShowStatsFPSStutter"),
-            localized_choice("OptionsGraphics", "ShowStatsFPSStutterTiming"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ValidationLayers,
-        label: lookup_key("OptionsGraphics", "ValidationLayers"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::HighDpi,
-        label: lookup_key("OptionsGraphics", "HighDPI"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::VisualDelay,
-        label: lookup_key("OptionsGraphics", "VisualDelay"),
-        choices: &[literal_choice("0 ms")],
-        inline: false,
-    },
-];
-
-pub const GRAPHICS_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::GfxVideoRenderer,
-        name: lookup_key("OptionsGraphics", "VideoRenderer"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "VideoRendererHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxSoftwareThreads,
-        name: lookup_key("OptionsGraphics", "SoftwareRendererThreads"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "SoftwareRendererThreadsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxDisplayMode,
-        name: lookup_key("OptionsGraphics", "DisplayMode"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "DisplayModeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxDisplayAspectRatio,
-        name: lookup_key("OptionsGraphics", "DisplayAspectRatio"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "DisplayAspectRatioHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxDisplayResolution,
-        name: lookup_key("OptionsGraphics", "DisplayResolution"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "DisplayResolutionHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxRefreshRate,
-        name: lookup_key("OptionsGraphics", "RefreshRate"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "RefreshRateHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxFullscreenType,
-        name: lookup_key("OptionsGraphics", "FullscreenType"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "FullscreenTypeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxVSync,
-        name: lookup_key("OptionsGraphics", "VSync"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "VSyncHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxPresentMode,
-        name: lookup_key("OptionsGraphics", "PresentMode"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "PresentModeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxMaxFps,
-        name: lookup_key("OptionsGraphics", "MaxFps"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "MaxFpsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxMaxFpsValue,
-        name: lookup_key("OptionsGraphics", "MaxFpsValue"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "MaxFpsValueHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxShowStats,
-        name: lookup_key("OptionsGraphics", "ShowStats"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "ShowStatsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxValidationLayers,
-        name: lookup_key("OptionsGraphics", "ValidationLayers"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "ValidationLayersHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxHighDpi,
-        name: lookup_key("OptionsGraphics", "HighDPI"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "HighDPIHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GfxVisualDelay,
-        name: lookup_key("OptionsGraphics", "VisualDelay"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGraphicsHelp",
-            "VisualDelayHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const INPUT_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::ConfigureMappings,
-        label: lookup_key("OptionsInput", "ConfigureMappings"),
-        choices: &[localized_choice("Common", "Open")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::TestInput,
-        label: lookup_key("OptionsInput", "TestInput"),
-        choices: &[localized_choice("Common", "Open")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::InputOptions,
-        label: lookup_key("OptionsInput", "InputOptions"),
-        choices: &[localized_choice("Common", "Open")],
-        inline: false,
-    },
-];
-
-pub const INPUT_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::InpConfigureMappings,
-        name: lookup_key("OptionsInput", "ConfigureMappings"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "ConfigureMappingsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::InpTestInput,
-        name: lookup_key("OptionsInput", "TestInput"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "TestInputHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::InpInputOptions,
-        name: lookup_key("OptionsInput", "InputOptions"),
-        help: &[
-            HelpEntry::Paragraph(lookup_key("OptionsInputHelp", "InputOptionsHelp")),
-            HelpEntry::Bullet(lookup_key("OptionsInput", "GamepadBackend")),
-            HelpEntry::Bullet(lookup_key("OptionsInput", "UseFSRs")),
-            HelpEntry::Bullet(lookup_key("OptionsInput", "MenuNavigation")),
-            HelpEntry::Bullet(lookup_key("OptionsInput", "OptionsNavigation")),
-            HelpEntry::Bullet(lookup_key("OptionsInput", "MenuButtons")),
-            HelpEntry::Bullet(lookup_key("OptionsInput", "Debounce")),
-        ],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const INPUT_BACKEND_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::GamepadBackend,
-        label: lookup_key("OptionsInput", "GamepadBackend"),
-        choices: INPUT_BACKEND_CHOICES,
-        inline: INPUT_BACKEND_INLINE,
-    },
-    SubRow {
-        id: SubRowId::UseFsrs,
-        label: lookup_key("OptionsInput", "UseFSRs"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MenuNavigation,
-        label: lookup_key("OptionsInput", "MenuNavigation"),
-        choices: &[
-            localized_choice("OptionsInput", "MenuNavigationFiveKey"),
-            localized_choice("OptionsInput", "MenuNavigationThreeKey"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::OptionsNavigation,
-        label: lookup_key("OptionsInput", "OptionsNavigation"),
-        choices: &[
-            localized_choice("OptionsInput", "OptionsNavigationStepMania"),
-            localized_choice("OptionsInput", "OptionsNavigationArcade"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MenuButtons,
-        label: lookup_key("OptionsInput", "MenuButtons"),
-        choices: &[
-            localized_choice("OptionsInput", "DedicatedMenuButtonsGameplay"),
-            localized_choice("OptionsInput", "DedicatedMenuButtonsOnly"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::Debounce,
-        label: lookup_key("OptionsInput", "Debounce"),
-        choices: &[literal_choice("20ms")],
-        inline: true,
-    },
-];
-
-pub const INPUT_BACKEND_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::InpGamepadBackend,
-        name: lookup_key("OptionsInput", "GamepadBackend"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "GamepadBackendHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::InpUseFsrs,
-        name: lookup_key("OptionsInput", "UseFSRs"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "UseFSRsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::InpMenuButtons,
-        name: lookup_key("OptionsInput", "MenuButtons"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "MenuButtonsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::InpOptionsNavigation,
-        name: lookup_key("OptionsInput", "OptionsNavigation"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "OptionsNavigationHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::InpMenuNavigation,
-        name: lookup_key("OptionsInput", "MenuNavigation"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "MenuNavigationHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::InpDebounce,
-        name: lookup_key("OptionsInput", "Debounce"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsInputHelp",
-            "DebounceHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const MACHINE_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::SelectProfile,
-        label: lookup_key("OptionsMachine", "SelectProfile"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::SelectColor,
-        label: lookup_key("OptionsMachine", "SelectColor"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::SelectStyle,
-        label: lookup_key("OptionsMachine", "SelectStyle"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::PreferredStyle,
-        label: lookup_key("OptionsMachine", "PreferredStyle"),
-        choices: &[
-            localized_choice("OptionsMachine", "PreferredStyleSingle"),
-            localized_choice("OptionsMachine", "PreferredStyleVersus"),
-            localized_choice("OptionsMachine", "PreferredStyleDouble"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::SelectPlayMode,
-        label: lookup_key("OptionsMachine", "SelectPlayMode"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::PreferredMode,
-        label: lookup_key("OptionsMachine", "PreferredMode"),
-        choices: &[
-            localized_choice("OptionsMachine", "PreferredModeRegular"),
-            localized_choice("OptionsMachine", "PreferredModeMarathon"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::Font,
-        label: lookup_key("OptionsMachine", "MachineFont"),
-        choices: &[
-            localized_choice("OptionsMachine", "MachineFontCommon"),
-            localized_choice("OptionsMachine", "MachineFontMega"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::EvalSummary,
-        label: lookup_key("OptionsMachine", "EvalSummary"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::NameEntry,
-        label: lookup_key("OptionsMachine", "NameEntry"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::GameoverScreen,
-        label: lookup_key("OptionsMachine", "GameoverScreen"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::WriteCurrentScreen,
-        label: lookup_key("OptionsMachine", "WriteCurrentScreen"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MenuMusic,
-        label: lookup_key("OptionsMachine", "MenuMusic"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MenuBackground,
-        label: lookup_key("OptionsMachine", "MenuBackground"),
-        choices: &[literal_choice("❤"), literal_choice("🌀")],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::Replays,
-        label: lookup_key("OptionsMachine", "Replays"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::PerPlayerGlobalOffsets,
-        label: lookup_key("OptionsMachine", "PerPlayerGlobalOffsets"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::KeyboardFeatures,
-        label: lookup_key("OptionsMachine", "KeyboardFeatures"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::VideoBgs,
-        label: lookup_key("OptionsMachine", "VideoBGs"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-];
-
-pub const MACHINE_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::MchSelectProfile,
-        name: lookup_key("OptionsMachine", "SelectProfile"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "SelectProfileHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchSelectColor,
-        name: lookup_key("OptionsMachine", "SelectColor"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "SelectColorHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchSelectStyle,
-        name: lookup_key("OptionsMachine", "SelectStyle"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "SelectStyleHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchPreferredStyle,
-        name: lookup_key("OptionsMachine", "PreferredStyle"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "PreferredStyleHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchSelectPlayMode,
-        name: lookup_key("OptionsMachine", "SelectPlayMode"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "SelectPlayModeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchPreferredMode,
-        name: lookup_key("OptionsMachine", "PreferredMode"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "PreferredModeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchFont,
-        name: lookup_key("OptionsMachine", "MachineFont"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "MachineFontHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchEvalSummary,
-        name: lookup_key("OptionsMachine", "EvalSummary"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "EvalSummaryHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchNameEntry,
-        name: lookup_key("OptionsMachine", "NameEntry"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "NameEntryHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchGameoverScreen,
-        name: lookup_key("OptionsMachine", "GameoverScreen"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "GameoverScreenHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchWriteCurrentScreen,
-        name: lookup_key("OptionsMachine", "WriteCurrentScreen"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "WriteCurrentScreenHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchMenuMusic,
-        name: lookup_key("OptionsMachine", "MenuMusic"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "MenuMusicHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchMenuBackground,
-        name: lookup_key("OptionsMachine", "MenuBackground"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "MenuBackgroundHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchReplays,
-        name: lookup_key("OptionsMachine", "Replays"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "ReplaysHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchPerPlayerGlobalOffsets,
-        name: lookup_key("OptionsMachine", "PerPlayerGlobalOffsets"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "PerPlayerGlobalOffsetsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchKeyboardFeatures,
-        name: lookup_key("OptionsMachine", "KeyboardFeatures"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "KeyboardFeaturesHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::MchVideoBgs,
-        name: lookup_key("OptionsMachine", "VideoBGs"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsMachineHelp",
-            "VideoBgsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const COURSE_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::ShowRandomCourses,
-        label: lookup_key("OptionsCourse", "ShowRandomCourses"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowMostPlayed,
-        label: lookup_key("OptionsCourse", "ShowMostPlayed"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowIndividualScores,
-        label: lookup_key("OptionsCourse", "ShowIndividualScores"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::AutosubmitIndividual,
-        label: lookup_key("OptionsCourse", "AutosubmitIndividual"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-];
-
-pub const COURSE_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::CrsShowRandom,
-        name: lookup_key("OptionsCourse", "ShowRandomCourses"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsCourseHelp",
-            "ShowRandomCoursesHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::CrsShowMostPlayed,
-        name: lookup_key("OptionsCourse", "ShowMostPlayed"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsCourseHelp",
-            "ShowMostPlayedHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::CrsShowIndividualScores,
-        name: lookup_key("OptionsCourse", "ShowIndividualScores"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsCourseHelp",
-            "ShowIndividualScoresHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::CrsAutosubmitIndividual,
-        name: lookup_key("OptionsCourse", "AutosubmitIndividual"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsCourseHelp",
-            "AutosubmitIndividualHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const GAMEPLAY_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::BgBrightness,
-        label: lookup_key("OptionsGameplay", "BGBrightness"),
-        choices: &[
-            literal_choice("0%"),
-            literal_choice("10%"),
-            literal_choice("20%"),
-            literal_choice("30%"),
-            literal_choice("40%"),
-            literal_choice("50%"),
-            literal_choice("60%"),
-            literal_choice("70%"),
-            literal_choice("80%"),
-            literal_choice("90%"),
-            literal_choice("100%"),
-        ],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::CenteredP1Notefield,
-        label: lookup_key("OptionsGameplay", "CenteredP1Notefield"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ZmodRatingBox,
-        label: lookup_key("OptionsGameplay", "ZmodRatingBox"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::BpmDecimal,
-        label: lookup_key("OptionsGameplay", "BpmDecimal"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::AutoScreenshot,
-        label: lookup_key("OptionsGameplay", "AutoScreenshot"),
-        choices: &[
-            literal_choice("PBs"),
-            literal_choice("Fails"),
-            literal_choice("Clears"),
-            literal_choice("Quads"),
-            literal_choice("Quints"),
-        ],
-        inline: true,
-    },
-];
-
-pub const GAMEPLAY_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::GpBgBrightness,
-        name: lookup_key("OptionsGameplay", "BGBrightness"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGameplayHelp",
-            "BgBrightnessHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GpCenteredP1,
-        name: lookup_key("OptionsGameplay", "CenteredP1Notefield"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGameplayHelp",
-            "CenteredP1NotefieldHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GpZmodRatingBox,
-        name: lookup_key("OptionsGameplay", "ZmodRatingBox"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGameplayHelp",
-            "ZmodRatingBoxHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GpBpmDecimal,
-        name: lookup_key("OptionsGameplay", "BpmDecimal"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGameplayHelp",
-            "BpmDecimalHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GpAutoScreenshot,
-        name: lookup_key("OptionsGameplay", "AutoScreenshot"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGameplayHelp",
-            "AutoScreenshotHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const SOUND_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::SoundDevice,
-        label: lookup_key("OptionsSound", "SoundDevice"),
-        choices: &[localized_choice("Common", "Auto")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::AudioOutputMode,
-        label: lookup_key("OptionsSound", "AudioOutputMode"),
-        choices: &[
-            localized_choice("OptionsSound", "OutputModeAuto"),
-            localized_choice("OptionsSound", "OutputModeShared"),
-        ],
-        inline: false,
-    },
-    #[cfg(target_os = "linux")]
-    SubRow {
-        id: SubRowId::LinuxAudioBackend,
-        label: lookup_key("OptionsSound", "LinuxAudioBackend"),
-        choices: SOUND_LINUX_BACKEND_CHOICES,
-        inline: false,
-    },
-    #[cfg(target_os = "linux")]
-    SubRow {
-        id: SubRowId::AlsaExclusive,
-        label: lookup_key("OptionsSound", "AlsaExclusive"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::AudioSampleRate,
-        label: lookup_key("OptionsSound", "AudioSampleRate"),
-        choices: &[localized_choice("Common", "Auto")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::MasterVolume,
-        label: lookup_key("OptionsSound", "MasterVolume"),
-        choices: &[literal_choice("100%")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::SfxVolume,
-        label: lookup_key("OptionsSound", "SFXVolume"),
-        choices: &[literal_choice("100%")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::AssistTickVolume,
-        label: lookup_key("OptionsSound", "AssistTickVolume"),
-        choices: &[literal_choice("100%")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::MusicVolume,
-        label: lookup_key("OptionsSound", "MusicVolume"),
-        choices: &[literal_choice("100%")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::MineSounds,
-        label: lookup_key("OptionsSound", "MineSounds"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::GlobalOffset,
-        label: lookup_key("OptionsSound", "GlobalOffset"),
-        choices: &[literal_choice("0 ms")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::RateModPreservesPitch,
-        label: lookup_key("OptionsSound", "RateModPreservesPitch"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-];
-
-pub const SOUND_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::SndDevice,
-        name: lookup_key("OptionsSound", "SoundDevice"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "SoundDeviceHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndOutputMode,
-        name: lookup_key("OptionsSound", "AudioOutputMode"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "AudioOutputModeHelp",
-        ))],
-    },
-    #[cfg(target_os = "linux")]
-    Item {
-        id: ItemId::SndLinuxBackend,
-        name: lookup_key("OptionsSound", "LinuxAudioBackend"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "LinuxAudioBackendHelp",
-        ))],
-    },
-    #[cfg(target_os = "linux")]
-    Item {
-        id: ItemId::SndAlsaExclusive,
-        name: lookup_key("OptionsSound", "AlsaExclusive"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "AlsaExclusiveHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndSampleRate,
-        name: lookup_key("OptionsSound", "AudioSampleRate"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "AudioSampleRateHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndMasterVolume,
-        name: lookup_key("OptionsSound", "MasterVolume"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "MasterVolumeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndSfxVolume,
-        name: lookup_key("OptionsSound", "SFXVolume"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "SfxVolumeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndAssistTickVolume,
-        name: lookup_key("OptionsSound", "AssistTickVolume"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "AssistTickVolumeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndMusicVolume,
-        name: lookup_key("OptionsSound", "MusicVolume"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "MusicVolumeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndMineSounds,
-        name: lookup_key("OptionsSound", "MineSounds"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "MineSoundsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndGlobalOffset,
-        name: lookup_key("OptionsSound", "GlobalOffset"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "GlobalOffsetHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SndRateModPitch,
-        name: lookup_key("OptionsSound", "RateModPreservesPitch"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSoundHelp",
-            "RateModPreservesPitchHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const SELECT_MUSIC_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::ShowBanners,
-        label: lookup_key("OptionsSelectMusic", "ShowBanners"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowVideoBanners,
-        label: lookup_key("OptionsSelectMusic", "ShowVideoBanners"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowBreakdown,
-        label: lookup_key("OptionsSelectMusic", "ShowBreakdown"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::BreakdownStyle,
-        label: lookup_key("OptionsSelectMusic", "BreakdownStyle"),
-        choices: &[literal_choice("SL"), literal_choice("SN")],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowNativeLanguage,
-        label: lookup_key("OptionsSelectMusic", "ShowNativeLanguage"),
-        choices: &[
-            localized_choice("OptionsSelectMusic", "NativeLanguageTranslit"),
-            localized_choice("OptionsSelectMusic", "NativeLanguageNative"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MusicWheelSpeed,
-        label: lookup_key("OptionsSelectMusic", "MusicWheelSpeed"),
-        choices: &[
-            localized_choice("OptionsSelectMusic", "WheelSpeedSlow"),
-            localized_choice("OptionsSelectMusic", "WheelSpeedNormal"),
-            localized_choice("OptionsSelectMusic", "WheelSpeedFast"),
-            localized_choice("OptionsSelectMusic", "WheelSpeedFaster"),
-            localized_choice("OptionsSelectMusic", "WheelSpeedRidiculous"),
-            localized_choice("OptionsSelectMusic", "WheelSpeedLudicrous"),
-            localized_choice("OptionsSelectMusic", "WheelSpeedPlaid"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MusicWheelStyle,
-        label: lookup_key("OptionsSelectMusic", "MusicWheelStyle"),
-        choices: &[literal_choice("ITG"), literal_choice("IIDX")],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowCdTitles,
-        label: lookup_key("OptionsSelectMusic", "ShowCDTitles"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowWheelGrades,
-        label: lookup_key("OptionsSelectMusic", "ShowWheelGrades"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowWheelLamps,
-        label: lookup_key("OptionsSelectMusic", "ShowWheelLamps"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ItlRank,
-        label: lookup_key("OptionsSelectMusic", "ITLRank"),
-        choices: &[
-            localized_choice("OptionsSelectMusic", "ItlRankNone"),
-            localized_choice("OptionsSelectMusic", "ItlRankChart"),
-            localized_choice("OptionsSelectMusic", "ItlRankOverall"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ItlWheelData,
-        label: lookup_key("OptionsSelectMusic", "ITLWheelData"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("OptionsSelectMusic", "ItlWheelScore"),
-            localized_choice("OptionsSelectMusic", "ItlWheelPointsScore"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::NewPackBadge,
-        label: lookup_key("OptionsSelectMusic", "NewPackBadge"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("OptionsSelectMusic", "NewPackOpenPack"),
-            localized_choice("OptionsSelectMusic", "NewPackHasScore"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowPatternInfo,
-        label: lookup_key("OptionsSelectMusic", "ShowPatternInfo"),
-        choices: &[
-            localized_choice("Common", "Auto"),
-            localized_choice("OptionsSelectMusic", "PatternInfoTech"),
-            localized_choice("OptionsSelectMusic", "PatternInfoStamina"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ChartInfo,
-        label: lookup_key("OptionsSelectMusic", "ChartInfo"),
-        choices: &[
-            localized_choice("OptionsSelectMusic", "ChartInfoPeakNPS"),
-            localized_choice("OptionsSelectMusic", "ChartInfoMatrixRating"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::MusicPreviews,
-        label: lookup_key("OptionsSelectMusic", "MusicPreviews"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::PreviewMarker,
-        label: lookup_key("OptionsSelectMusic", "PreviewMarker"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::LoopMusic,
-        label: lookup_key("OptionsSelectMusic", "LoopMusic"),
-        choices: &[
-            localized_choice("OptionsSelectMusic", "LoopMusicPlayOnce"),
-            localized_choice("OptionsSelectMusic", "LoopMusicLoop"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowGameplayTimer,
-        label: lookup_key("OptionsSelectMusic", "ShowGameplayTimer"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ShowGsBox,
-        label: lookup_key("OptionsSelectMusic", "ShowGSBox"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::GsBoxPlacement,
-        label: lookup_key("OptionsSelectMusic", "GSBoxPlacement"),
-        choices: &[
-            localized_choice("Common", "Auto"),
-            localized_choice("OptionsSelectMusic", "GsBoxStepPane"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::GsBoxLeaderboards,
-        label: lookup_key("OptionsSelectMusic", "GSBoxLeaderboards"),
-        choices: &[
-            localized_choice("OptionsSelectMusic", "ScoreboxCycleITG"),
-            localized_choice("OptionsSelectMusic", "ScoreboxCycleEX"),
-            localized_choice("OptionsSelectMusic", "ScoreboxCycleHEX"),
-            localized_choice("OptionsSelectMusic", "ScoreboxCycleTournaments"),
-        ],
-        inline: true,
-    },
-];
-
-pub const SELECT_MUSIC_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::SmShowBanners,
-        name: lookup_key("OptionsSelectMusic", "ShowBanners"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowBannersHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmShowVideoBanners,
-        name: lookup_key("OptionsSelectMusic", "ShowVideoBanners"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowVideoBannersHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmShowBreakdown,
-        name: lookup_key("OptionsSelectMusic", "ShowBreakdown"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowBreakdownHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmBreakdownStyle,
-        name: lookup_key("OptionsSelectMusic", "BreakdownStyle"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "BreakdownStyleHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmNativeLanguage,
-        name: lookup_key("OptionsSelectMusic", "ShowNativeLanguage"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowNativeLanguageHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmWheelSpeed,
-        name: lookup_key("OptionsSelectMusic", "MusicWheelSpeed"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "MusicWheelSpeedHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmWheelStyle,
-        name: lookup_key("OptionsSelectMusic", "MusicWheelStyle"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "MusicWheelStyleHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmCdTitles,
-        name: lookup_key("OptionsSelectMusic", "ShowCDTitles"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowCdTitlesHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmWheelGrades,
-        name: lookup_key("OptionsSelectMusic", "ShowWheelGrades"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowWheelGradesHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmWheelLamps,
-        name: lookup_key("OptionsSelectMusic", "ShowWheelLamps"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowWheelLampsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmWheelItlRank,
-        name: lookup_key("OptionsSelectMusic", "ITLRank"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ITLRankHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmWheelItl,
-        name: lookup_key("OptionsSelectMusic", "ITLWheelData"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ItlWheelDataHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmNewPackBadge,
-        name: lookup_key("OptionsSelectMusic", "NewPackBadge"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "NewPackBadgeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmPatternInfo,
-        name: lookup_key("OptionsSelectMusic", "ShowPatternInfo"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowPatternInfoHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmChartInfo,
-        name: lookup_key("OptionsSelectMusic", "ChartInfo"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ChartInfoHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmPreviews,
-        name: lookup_key("OptionsSelectMusic", "MusicPreviews"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "MusicPreviewsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmPreviewMarker,
-        name: lookup_key("OptionsSelectMusic", "PreviewMarker"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "PreviewMarkerHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmPreviewLoop,
-        name: lookup_key("OptionsSelectMusic", "LoopMusic"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "LoopMusicHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmGameplayTimer,
-        name: lookup_key("OptionsSelectMusic", "ShowGameplayTimer"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowGameplayTimerHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmShowRivals,
-        name: lookup_key("OptionsSelectMusic", "ShowGSBox"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "ShowGsBoxHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmScoreboxPlacement,
-        name: lookup_key("OptionsSelectMusic", "GSBoxPlacement"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "GsBoxPlacementHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SmScoreboxCycle,
-        name: lookup_key("OptionsSelectMusic", "GSBoxLeaderboards"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSelectMusicHelp",
-            "GsBoxLeaderboardsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const ADVANCED_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::DefaultFailType,
-        label: lookup_key("OptionsAdvanced", "DefaultFailType"),
-        choices: &[
-            localized_choice("OptionsAdvanced", "FailImmediate"),
-            localized_choice("OptionsAdvanced", "FailImmediateContinue"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::BannerCache,
-        label: lookup_key("OptionsAdvanced", "BannerCache"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::CdTitleCache,
-        label: lookup_key("OptionsAdvanced", "CDTitleCache"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::SongParsingThreads,
-        label: lookup_key("OptionsAdvanced", "SongParsingThreads"),
-        choices: &[localized_choice("Common", "Auto")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::CacheSongs,
-        label: lookup_key("OptionsAdvanced", "CacheSongs"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::FastLoad,
-        label: lookup_key("OptionsAdvanced", "FastLoad"),
-        choices: &[
-            localized_choice("Common", "Off"),
-            localized_choice("Common", "On"),
-        ],
-        inline: true,
-    },
-];
-
-pub const ADVANCED_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::AdvDefaultFailType,
-        name: lookup_key("OptionsAdvanced", "DefaultFailType"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsAdvancedHelp",
-            "DefaultFailTypeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::AdvBannerCache,
-        name: lookup_key("OptionsAdvanced", "BannerCache"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsAdvancedHelp",
-            "BannerCacheHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::AdvCdTitleCache,
-        name: lookup_key("OptionsAdvanced", "CDTitleCache"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsAdvancedHelp",
-            "CdTitleCacheHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::AdvSongParsingThreads,
-        name: lookup_key("OptionsAdvanced", "SongParsingThreads"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsAdvancedHelp",
-            "SongParsingThreadsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::AdvCacheSongs,
-        name: lookup_key("OptionsAdvanced", "CacheSongs"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsAdvancedHelp",
-            "CacheSongsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::AdvFastLoad,
-        name: lookup_key("OptionsAdvanced", "FastLoad"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsAdvancedHelp",
-            "FastLoadHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const GROOVESTATS_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::EnableGrooveStats,
-        label: lookup_key("OptionsGrooveStats", "EnableGrooveStats"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::EnableBoogieStats,
-        label: lookup_key("OptionsGrooveStats", "EnableBoogieStats"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::GsSubmitFails,
-        label: lookup_key("OptionsGrooveStats", "GsSubmitFails"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::AutoPopulateScores,
-        label: lookup_key("OptionsGrooveStats", "AutoPopulateScores"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::AutoDownloadUnlocks,
-        label: lookup_key("OptionsGrooveStats", "AutoDownloadUnlocks"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::SeparateUnlocksByPlayer,
-        label: lookup_key("OptionsGrooveStats", "SeparateUnlocksByPlayer"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-];
-
-pub const ARROWCLOUD_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::EnableArrowCloud,
-        label: lookup_key("OptionsGrooveStats", "EnableArrowCloud"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ArrowCloudSubmitFails,
-        label: lookup_key("OptionsGrooveStats", "ArrowCloudSubmitFails"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-];
-
-pub const ONLINE_SCORING_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::GsBsOptions,
-        label: lookup_key("OptionsOnlineScoring", "GsBsOptions"),
-        choices: &[],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::ArrowCloudOptions,
-        label: lookup_key("OptionsOnlineScoring", "ArrowCloudOptions"),
-        choices: &[],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::ScoreImport,
-        label: lookup_key("OptionsOnlineScoring", "ScoreImport"),
-        choices: &[],
-        inline: false,
-    },
-];
-
-pub const NULL_OR_DIE_MENU_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::NullOrDieOptions,
-        label: lookup_key("OptionsOnlineScoring", "NullOrDieOptions"),
-        choices: &[],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::SyncPacks,
-        label: lookup_key("OptionsOnlineScoring", "SyncPacks"),
-        choices: &[],
-        inline: false,
-    },
-];
-
-pub const NULL_OR_DIE_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::SyncGraph,
-        label: lookup_key("OptionsNullOrDie", "SyncGraph"),
-        choices: &[
-            localized_choice("OptionsNullOrDie", "SyncGraphFrequency"),
-            localized_choice("OptionsNullOrDie", "SyncGraphBeatIndex"),
-            localized_choice("OptionsNullOrDie", "SyncGraphPostKernelFingerprint"),
-        ],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::SyncConfidence,
-        label: lookup_key("OptionsNullOrDie", "SyncConfidence"),
-        choices: &[
-            literal_choice("0%"),
-            literal_choice("5%"),
-            literal_choice("10%"),
-            literal_choice("15%"),
-            literal_choice("20%"),
-            literal_choice("25%"),
-            literal_choice("30%"),
-            literal_choice("35%"),
-            literal_choice("40%"),
-            literal_choice("45%"),
-            literal_choice("50%"),
-            literal_choice("55%"),
-            literal_choice("60%"),
-            literal_choice("65%"),
-            literal_choice("70%"),
-            literal_choice("75%"),
-            literal_choice("80%"),
-            literal_choice("85%"),
-            literal_choice("90%"),
-            literal_choice("95%"),
-            literal_choice("100%"),
-        ],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::PackSyncThreads,
-        label: lookup_key("OptionsNullOrDie", "PackSyncThreads"),
-        choices: &[localized_choice("Common", "Auto")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::Fingerprint,
-        label: lookup_key("OptionsNullOrDie", "Fingerprint"),
-        choices: &[literal_choice("50.0 ms")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::Window,
-        label: lookup_key("OptionsNullOrDie", "Window"),
-        choices: &[literal_choice("10.0 ms")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::Step,
-        label: lookup_key("OptionsNullOrDie", "Step"),
-        choices: &[literal_choice("0.2 ms")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::MagicOffset,
-        label: lookup_key("OptionsNullOrDie", "MagicOffset"),
-        choices: &[literal_choice("0.0 ms")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::KernelTarget,
-        label: lookup_key("OptionsNullOrDie", "KernelTarget"),
-        choices: &[
-            localized_choice("OptionsNullOrDie", "KernelTargetDigest"),
-            localized_choice("OptionsNullOrDie", "KernelTargetAccumulator"),
-        ],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::KernelType,
-        label: lookup_key("OptionsNullOrDie", "KernelType"),
-        choices: &[
-            localized_choice("OptionsNullOrDie", "KernelTypeRising"),
-            localized_choice("OptionsNullOrDie", "KernelTypeLoudest"),
-        ],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::FullSpectrogram,
-        label: lookup_key("OptionsNullOrDie", "FullSpectrogram"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: false,
-    },
-];
-
-pub const SYNC_PACK_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::SyncPackPack,
-        label: lookup_key("OptionsSyncPack", "SyncPackPack"),
-        choices: &[localized_choice("OptionsSyncPack", "AllPacks")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::SyncPackStart,
-        label: lookup_key("OptionsSyncPack", "SyncPackStart"),
-        choices: &[localized_choice("Common", "Start")],
-        inline: false,
-    },
-];
-
-pub const SCORE_IMPORT_OPTIONS_ROWS: &[SubRow] = &[
-    SubRow {
-        id: SubRowId::ScoreImportEndpoint,
-        label: lookup_key("OptionsScoreImport", "ScoreImportEndpoint"),
-        choices: &[
-            literal_choice("GrooveStats"),
-            literal_choice("BoogieStats"),
-            literal_choice("ArrowCloud"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ScoreImportProfile,
-        label: lookup_key("OptionsScoreImport", "ScoreImportProfile"),
-        choices: &[localized_choice("OptionsScoreImport", "NoEligibleProfiles")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::ScoreImportPack,
-        label: lookup_key("OptionsScoreImport", "ScoreImportPack"),
-        choices: &[localized_choice("OptionsScoreImport", "AllPacks")],
-        inline: false,
-    },
-    SubRow {
-        id: SubRowId::ScoreImportOnlyMissing,
-        label: lookup_key("OptionsScoreImport", "ScoreImportOnlyMissing"),
-        choices: &[
-            localized_choice("Common", "No"),
-            localized_choice("Common", "Yes"),
-        ],
-        inline: true,
-    },
-    SubRow {
-        id: SubRowId::ScoreImportStart,
-        label: lookup_key("OptionsScoreImport", "ScoreImportStart"),
-        choices: &[localized_choice("Common", "Start")],
-        inline: false,
-    },
-];
-
-pub const GROOVESTATS_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::GsEnable,
-        name: lookup_key("OptionsGrooveStats", "EnableGrooveStats"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "EnableGrooveStatsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GsEnableBoogie,
-        name: lookup_key("OptionsGrooveStats", "EnableBoogieStats"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "EnableBoogieStatsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GsSubmitFails,
-        name: lookup_key("OptionsGrooveStats", "GsSubmitFails"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "GsSubmitFailsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GsAutoPopulate,
-        name: lookup_key("OptionsGrooveStats", "AutoPopulateScores"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "AutoPopulateScoresHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GsAutoDownloadUnlocks,
-        name: lookup_key("OptionsGrooveStats", "AutoDownloadUnlocks"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "AutoDownloadUnlocksHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::GsSeparateUnlocks,
-        name: lookup_key("OptionsGrooveStats", "SeparateUnlocksByPlayer"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "SeparateUnlocksByPlayerHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const ARROWCLOUD_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::AcEnable,
-        name: lookup_key("OptionsGrooveStats", "EnableArrowCloud"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "EnableArrowCloudHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::AcSubmitFails,
-        name: lookup_key("OptionsGrooveStats", "ArrowCloudSubmitFails"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsGrooveStatsHelp",
-            "ArrowCloudSubmitFailsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const ONLINE_SCORING_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::OsGsBsOptions,
-        name: lookup_key("OptionsOnlineScoring", "GsBsOptions"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsOnlineScoringHelp",
-            "GsBsOptionsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::OsArrowCloudOptions,
-        name: lookup_key("OptionsOnlineScoring", "ArrowCloudOptions"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsOnlineScoringHelp",
-            "ArrowCloudOptionsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::OsScoreImport,
-        name: lookup_key("OptionsOnlineScoring", "ScoreImport"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsOnlineScoringHelp",
-            "ScoreImportHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const NULL_OR_DIE_MENU_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::NodOptions,
-        name: lookup_key("OptionsOnlineScoring", "NullOrDieOptions"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsOnlineScoringHelp",
-            "NullOrDieOptionsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodSyncPacks,
-        name: lookup_key("OptionsOnlineScoring", "SyncPacks"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsOnlineScoringHelp",
-            "SyncPacksHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const NULL_OR_DIE_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::NodSyncGraph,
-        name: lookup_key("OptionsNullOrDie", "SyncGraph"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "SyncGraphHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodSyncConfidence,
-        name: lookup_key("OptionsNullOrDie", "SyncConfidence"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "SyncConfidenceHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodPackSyncThreads,
-        name: lookup_key("OptionsNullOrDie", "PackSyncThreads"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "PackSyncThreadsHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodFingerprint,
-        name: lookup_key("OptionsNullOrDie", "Fingerprint"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "FingerprintHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodWindow,
-        name: lookup_key("OptionsNullOrDie", "Window"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "WindowHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodStep,
-        name: lookup_key("OptionsNullOrDie", "Step"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "StepHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodMagicOffset,
-        name: lookup_key("OptionsNullOrDie", "MagicOffset"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "MagicOffsetHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodKernelTarget,
-        name: lookup_key("OptionsNullOrDie", "KernelTarget"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "KernelTargetHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodKernelType,
-        name: lookup_key("OptionsNullOrDie", "KernelType"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "KernelTypeHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::NodFullSpectrogram,
-        name: lookup_key("OptionsNullOrDie", "FullSpectrogram"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsNullOrDieHelp",
-            "FullSpectrogramHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const SYNC_PACK_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::SpPack,
-        name: lookup_key("OptionsSyncPack", "SyncPackPack"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSyncPackHelp",
-            "SyncPackPackHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SpStart,
-        name: lookup_key("OptionsSyncPack", "SyncPackStart"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsSyncPackHelp",
-            "SyncPackStartHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
-
-pub const SCORE_IMPORT_OPTIONS_ITEMS: &[Item] = &[
-    Item {
-        id: ItemId::SiEndpoint,
-        name: lookup_key("OptionsScoreImport", "ScoreImportEndpoint"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsScoreImportHelp",
-            "ScoreImportEndpointHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SiProfile,
-        name: lookup_key("OptionsScoreImport", "ScoreImportProfile"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsScoreImportHelp",
-            "ScoreImportProfileHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SiPack,
-        name: lookup_key("OptionsScoreImport", "ScoreImportPack"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsScoreImportHelp",
-            "ScoreImportPackHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SiOnlyMissing,
-        name: lookup_key("OptionsScoreImport", "ScoreImportOnlyMissing"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsScoreImportHelp",
-            "ScoreImportOnlyMissingHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::SiStart,
-        name: lookup_key("OptionsScoreImport", "ScoreImportStart"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsScoreImportHelp",
-            "ScoreImportStartHelp",
-        ))],
-    },
-    Item {
-        id: ItemId::Exit,
-        name: lookup_key("Options", "Exit"),
-        help: &[HelpEntry::Paragraph(lookup_key(
-            "OptionsHelp",
-            "ExitSubHelp",
-        ))],
-    },
-];
 
 /// Returns `true` when the given submenu row should be treated as disabled
 /// (non-interactive and visually dimmed). Add new cases here for any row
@@ -3694,218 +1154,6 @@ const fn submenu_title(kind: SubmenuKind) -> &'static str {
         SubmenuKind::ArrowCloud => "ARROWCLOUD OPTIONS",
         SubmenuKind::ScoreImport => "SCORE IMPORT",
     }
-}
-
-fn backend_to_renderer_choice_index(backend: BackendType) -> usize {
-    VIDEO_RENDERER_OPTIONS
-        .iter()
-        .position(|(b, _)| *b == backend)
-        .unwrap_or(0)
-}
-
-fn renderer_choice_index_to_backend(idx: usize) -> BackendType {
-    VIDEO_RENDERER_OPTIONS
-        .get(idx)
-        .map_or_else(|| VIDEO_RENDERER_OPTIONS[0].0, |(backend, _)| *backend)
-}
-
-fn selected_video_renderer(state: &State) -> BackendType {
-    let choice_idx = state
-        .sub_choice_indices_graphics
-        .get(VIDEO_RENDERER_ROW_INDEX)
-        .copied()
-        .unwrap_or(0);
-    renderer_choice_index_to_backend(choice_idx)
-}
-
-fn build_software_thread_choices() -> Vec<u8> {
-    let max_threads = std::thread::available_parallelism()
-        .map(std::num::NonZero::get)
-        .unwrap_or(8)
-        .clamp(2, 32);
-    let mut out = Vec::with_capacity(max_threads + 1);
-    out.push(0); // Auto
-    for n in 1..=max_threads {
-        out.push(n as u8);
-    }
-    out
-}
-
-fn software_thread_choice_labels(values: &[u8]) -> Vec<String> {
-    values
-        .iter()
-        .map(|v| {
-            if *v == 0 {
-                tr("Common", "Auto").to_string()
-            } else {
-                v.to_string()
-            }
-        })
-        .collect()
-}
-
-fn software_thread_choice_index(values: &[u8], thread_count: u8) -> usize {
-    values
-        .iter()
-        .position(|&v| v == thread_count)
-        .unwrap_or_else(|| {
-            values
-                .iter()
-                .enumerate()
-                .min_by_key(|(_, v)| v.abs_diff(thread_count))
-                .map_or(0, |(idx, _)| idx)
-        })
-}
-
-fn software_thread_from_choice(values: &[u8], idx: usize) -> u8 {
-    values.get(idx).copied().unwrap_or(0)
-}
-
-fn build_max_fps_choices() -> Vec<u16> {
-    let mut out = Vec::with_capacity(
-        1 + usize::from(MAX_FPS_MAX.saturating_sub(MAX_FPS_MIN)) / usize::from(MAX_FPS_STEP),
-    );
-    let mut fps = MAX_FPS_MIN;
-    while fps <= MAX_FPS_MAX {
-        out.push(fps);
-        fps = fps.saturating_add(MAX_FPS_STEP);
-    }
-    out
-}
-
-fn max_fps_choice_labels(values: &[u16]) -> Vec<String> {
-    values.iter().map(ToString::to_string).collect()
-}
-
-#[inline(always)]
-const fn clamped_max_fps(max_fps: u16) -> u16 {
-    if max_fps < MAX_FPS_MIN {
-        MAX_FPS_MIN
-    } else if max_fps > MAX_FPS_MAX {
-        MAX_FPS_MAX
-    } else {
-        max_fps
-    }
-}
-
-fn max_fps_choice_index(values: &[u16], max_fps: u16) -> usize {
-    let target = clamped_max_fps(max_fps);
-    values.iter().position(|&v| v == target).unwrap_or_else(|| {
-        values
-            .iter()
-            .enumerate()
-            .min_by_key(|(_, v)| v.abs_diff(target))
-            .map_or(0, |(idx, _)| idx)
-    })
-}
-
-fn max_fps_from_choice(values: &[u16], idx: usize) -> u16 {
-    values.get(idx).copied().unwrap_or(MAX_FPS_DEFAULT)
-}
-
-#[inline(always)]
-const fn present_mode_choice_index(mode: PresentModePolicy) -> usize {
-    match mode {
-        PresentModePolicy::Mailbox => 0,
-        PresentModePolicy::Immediate => 1,
-    }
-}
-
-#[inline(always)]
-const fn present_mode_from_choice(idx: usize) -> PresentModePolicy {
-    match idx {
-        1 => PresentModePolicy::Immediate,
-        _ => PresentModePolicy::Mailbox,
-    }
-}
-
-fn selected_present_mode_policy(state: &State) -> PresentModePolicy {
-    state
-        .sub_choice_indices_graphics
-        .get(PRESENT_MODE_ROW_INDEX)
-        .copied()
-        .map_or(state.present_mode_policy_at_load, present_mode_from_choice)
-}
-
-fn selected_high_dpi(state: &State) -> bool {
-    GRAPHICS_OPTIONS_ROWS
-        .iter()
-        .position(|row| row.id == SubRowId::HighDpi)
-        .and_then(|idx| state.sub_choice_indices_graphics.get(idx).copied())
-        .is_some_and(yes_no_from_choice)
-}
-
-#[inline(always)]
-fn set_max_fps_enabled_choice(state: &mut State, enabled: bool) {
-    let idx = yes_no_choice_index(enabled);
-    if let Some(slot) = state
-        .sub_choice_indices_graphics
-        .get_mut(MAX_FPS_ENABLED_ROW_INDEX)
-    {
-        *slot = idx;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_graphics
-        .get_mut(MAX_FPS_ENABLED_ROW_INDEX)
-    {
-        *slot = idx;
-    }
-}
-
-#[inline(always)]
-fn set_max_fps_value_choice_index(state: &mut State, idx: usize) {
-    let max_idx = state.max_fps_choices.len().saturating_sub(1);
-    let clamped = idx.min(max_idx);
-    if let Some(slot) = state
-        .sub_choice_indices_graphics
-        .get_mut(MAX_FPS_VALUE_ROW_INDEX)
-    {
-        *slot = clamped;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_graphics
-        .get_mut(MAX_FPS_VALUE_ROW_INDEX)
-    {
-        *slot = clamped;
-    }
-}
-
-#[inline(always)]
-fn graphics_show_software_threads(state: &State) -> bool {
-    selected_video_renderer(state) == BackendType::Software
-}
-
-#[inline(always)]
-fn graphics_show_present_mode(state: &State) -> bool {
-    state
-        .sub_choice_indices_graphics
-        .get(VSYNC_ROW_INDEX)
-        .copied()
-        .is_some_and(|idx| !yes_no_from_choice(idx))
-}
-
-#[inline(always)]
-fn graphics_show_max_fps(state: &State) -> bool {
-    graphics_show_present_mode(state)
-}
-
-#[inline(always)]
-fn max_fps_enabled(state: &State) -> bool {
-    state
-        .sub_choice_indices_graphics
-        .get(MAX_FPS_ENABLED_ROW_INDEX)
-        .copied()
-        .is_some_and(yes_no_from_choice)
-}
-
-#[inline(always)]
-fn graphics_show_max_fps_value(state: &State) -> bool {
-    graphics_show_max_fps(state) && max_fps_enabled(state)
-}
-
-#[inline(always)]
-fn graphics_show_high_dpi(state: &State) -> bool {
-    cfg!(target_os = "macos") && selected_video_renderer(state) == BackendType::OpenGL
 }
 
 fn submenu_visible_row_indices(state: &State, kind: SubmenuKind, rows: &[SubRow]) -> Vec<usize> {
@@ -4051,421 +1299,6 @@ const fn windows_backend_from_choice(idx: usize) -> WindowsPadBackend {
         0 => WindowsPadBackend::RawInput,
         _ => WindowsPadBackend::Wgi,
     }
-}
-
-const fn fullscreen_type_to_choice_index(fullscreen_type: FullscreenType) -> usize {
-    match fullscreen_type {
-        FullscreenType::Exclusive => 0,
-        FullscreenType::Borderless => 1,
-    }
-}
-
-const fn choice_index_to_fullscreen_type(idx: usize) -> FullscreenType {
-    match idx {
-        1 => FullscreenType::Borderless,
-        _ => FullscreenType::Exclusive,
-    }
-}
-
-fn selected_fullscreen_type(state: &State) -> FullscreenType {
-    state
-        .sub_choice_indices_graphics
-        .get(FULLSCREEN_TYPE_ROW_INDEX)
-        .copied()
-        .map_or(FullscreenType::Exclusive, choice_index_to_fullscreen_type)
-}
-
-fn selected_display_mode(state: &State) -> DisplayMode {
-    let display_choice = state
-        .sub_choice_indices_graphics
-        .get(DISPLAY_MODE_ROW_INDEX)
-        .copied()
-        .unwrap_or(0);
-    let windowed_idx = state.display_mode_choices.len().saturating_sub(1);
-    if windowed_idx == 0 || display_choice >= windowed_idx {
-        DisplayMode::Windowed
-    } else {
-        DisplayMode::Fullscreen(selected_fullscreen_type(state))
-    }
-}
-
-fn selected_display_monitor(state: &State) -> usize {
-    let display_choice = state
-        .sub_choice_indices_graphics
-        .get(DISPLAY_MODE_ROW_INDEX)
-        .copied()
-        .unwrap_or(0);
-    let windowed_idx = state.display_mode_choices.len().saturating_sub(1);
-    if windowed_idx == 0 || display_choice >= windowed_idx {
-        0
-    } else {
-        display_choice.min(windowed_idx.saturating_sub(1))
-    }
-}
-
-fn selected_refresh_rate_millihertz(state: &State) -> u32 {
-    let idx = state
-        .sub_choice_indices_graphics
-        .get(REFRESH_RATE_ROW_INDEX)
-        .copied()
-        .unwrap_or(0);
-    state.refresh_rate_choices.get(idx).copied().unwrap_or(0)
-}
-
-fn max_fps_seed_value(state: &State, max_fps: u16) -> u16 {
-    if max_fps != 0 {
-        return clamped_max_fps(max_fps);
-    }
-
-    let selected_refresh_mhz = selected_refresh_rate_millihertz(state);
-    let refresh_mhz = if selected_refresh_mhz != 0 {
-        selected_refresh_mhz
-    } else if let Some(spec) = state.monitor_specs.get(selected_display_monitor(state)) {
-        if matches!(selected_display_mode(state), DisplayMode::Fullscreen(_)) {
-            let (width, height) = selected_resolution(state);
-            display::supported_refresh_rates(Some(spec), width, height)
-                .into_iter()
-                .max()
-                .or_else(|| {
-                    spec.modes
-                        .iter()
-                        .map(|mode| mode.refresh_rate_millihertz)
-                        .max()
-                })
-                .unwrap_or(60_000)
-        } else {
-            spec.modes
-                .iter()
-                .map(|mode| mode.refresh_rate_millihertz)
-                .max()
-                .unwrap_or(60_000)
-        }
-    } else {
-        60_000
-    };
-
-    clamped_max_fps(((refresh_mhz + 500) / 1000) as u16)
-}
-
-fn seed_max_fps_value_choice(state: &mut State, max_fps: u16) {
-    let seeded = max_fps_seed_value(state, max_fps);
-    let idx = max_fps_choice_index(&state.max_fps_choices, seeded);
-    set_max_fps_value_choice_index(state, idx);
-}
-
-fn selected_max_fps(state: &State) -> u16 {
-    if !max_fps_enabled(state) {
-        return 0;
-    }
-    let idx = state
-        .sub_choice_indices_graphics
-        .get(MAX_FPS_VALUE_ROW_INDEX)
-        .copied()
-        .unwrap_or(0);
-    max_fps_from_choice(&state.max_fps_choices, idx)
-}
-
-fn ensure_display_mode_choices(state: &mut State) {
-    state.display_mode_choices = build_display_mode_choices(&state.monitor_specs);
-    // If current selection is out of bounds, reset it.
-    if let Some(idx) = state
-        .sub_choice_indices_graphics
-        .get_mut(DISPLAY_MODE_ROW_INDEX)
-        && *idx >= state.display_mode_choices.len()
-    {
-        *idx = 0;
-    }
-    if let Some(choice_idx) = state
-        .sub_choice_indices_graphics
-        .get(DISPLAY_MODE_ROW_INDEX)
-        .copied()
-        && let Some(cursor_idx) = state
-            .sub_cursor_indices_graphics
-            .get_mut(DISPLAY_MODE_ROW_INDEX)
-    {
-        *cursor_idx = choice_idx;
-    }
-    // Also re-run logic that depends on the selected monitor.
-    let current_res = selected_resolution(state);
-    rebuild_resolution_choices(state, current_res.0, current_res.1);
-}
-
-pub fn update_monitor_specs(state: &mut State, specs: Vec<MonitorSpec>) {
-    state.monitor_specs = specs;
-    ensure_display_mode_choices(state);
-    // Keep the Display Mode row aligned with the actual current mode after monitors refresh.
-    set_display_mode_row_selection(
-        state,
-        state.monitor_specs.len(),
-        state.display_mode_at_load,
-        state.display_monitor_at_load,
-    );
-    if state.max_fps_at_load == 0 && !max_fps_enabled(state) {
-        seed_max_fps_value_choice(state, 0);
-    }
-    clear_render_cache(state);
-}
-
-fn set_display_mode_row_selection(
-    state: &mut State,
-    _monitor_count: usize, // Ignored, we use stored monitor_specs now
-    mode: DisplayMode,
-    monitor: usize,
-) {
-    // Ensure choices are up to date.
-    ensure_display_mode_choices(state);
-    let windowed_idx = state.display_mode_choices.len().saturating_sub(1);
-    let idx = match mode {
-        DisplayMode::Windowed => windowed_idx,
-        DisplayMode::Fullscreen(_) => {
-            let max_idx = windowed_idx.saturating_sub(1);
-            if max_idx == 0 {
-                0
-            } else {
-                monitor.min(max_idx)
-            }
-        }
-    };
-    if let Some(slot) = state
-        .sub_choice_indices_graphics
-        .get_mut(DISPLAY_MODE_ROW_INDEX)
-    {
-        *slot = idx;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_graphics
-        .get_mut(DISPLAY_MODE_ROW_INDEX)
-    {
-        *slot = idx;
-    }
-    // Re-trigger resolution rebuild based on the potentially new monitor selection.
-    let current_res = selected_resolution(state);
-    rebuild_resolution_choices(state, current_res.0, current_res.1);
-}
-
-fn selected_aspect_label(state: &State) -> &'static str {
-    let idx = state
-        .sub_choice_indices_graphics
-        .get(DISPLAY_ASPECT_RATIO_ROW_INDEX)
-        .copied()
-        .unwrap_or(0);
-    DISPLAY_ASPECT_RATIO_CHOICES
-        .get(idx)
-        .or(Some(&DISPLAY_ASPECT_RATIO_CHOICES[0]))
-        .and_then(|c| c.as_str_static())
-        .unwrap_or("16:9")
-}
-
-fn inferred_aspect_choice(width: u32, height: u32) -> usize {
-    if height == 0 {
-        return 0;
-    }
-
-    if let Some(idx) = DISPLAY_ASPECT_RATIO_CHOICES.iter().position(|c| {
-        c.as_str_static()
-            .map_or(false, |label| aspect_matches(width, height, label))
-    }) {
-        return idx;
-    }
-
-    let ratio = width as f32 / height as f32;
-    let mut best_idx = 0;
-    let mut best_delta = f32::INFINITY;
-    for (idx, choice) in DISPLAY_ASPECT_RATIO_CHOICES.iter().enumerate() {
-        let Some(label) = choice.as_str_static() else {
-            continue;
-        };
-        let target = match label {
-            "16:9" => 16.0 / 9.0,
-            "16:10" => 16.0 / 10.0,
-            "4:3" => 4.0 / 3.0,
-            "1:1" => 1.0,
-            _ => continue,
-        };
-        let delta = (ratio - target).abs();
-        if delta < best_delta {
-            best_delta = delta;
-            best_idx = idx;
-        }
-    }
-    best_idx
-}
-
-fn sync_display_aspect_ratio(state: &mut State, width: u32, height: u32) {
-    let idx = inferred_aspect_choice(width, height);
-    if let Some(slot) = state
-        .sub_choice_indices_graphics
-        .get_mut(DISPLAY_ASPECT_RATIO_ROW_INDEX)
-    {
-        *slot = idx;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_graphics
-        .get_mut(DISPLAY_ASPECT_RATIO_ROW_INDEX)
-    {
-        *slot = idx;
-    }
-}
-
-fn push_unique_resolution(target: &mut Vec<(u32, u32)>, width: u32, height: u32) {
-    if !target.iter().any(|&(w, h)| w == width && h == height) {
-        target.push((width, height));
-    }
-}
-
-fn preset_resolutions_for_aspect(label: &str) -> Vec<(u32, u32)> {
-    match label.to_ascii_lowercase().as_str() {
-        "16:9" => vec![(1280, 720), (1600, 900), (1920, 1080)],
-        "16:10" => vec![(1280, 800), (1440, 900), (1680, 1050), (1920, 1200)],
-        "4:3" => vec![
-            (640, 480),
-            (800, 600),
-            (1024, 768),
-            (1280, 960),
-            (1600, 1200),
-        ],
-        "1:1" => vec![(342, 342), (456, 456), (608, 608), (810, 810), (1080, 1080)],
-        _ => DEFAULT_RESOLUTION_CHOICES.to_vec(),
-    }
-}
-
-fn aspect_matches(width: u32, height: u32, label: &str) -> bool {
-    let ratio = width as f32 / height as f32;
-    match label {
-        "16:9" => (ratio - 1.7777).abs() < 0.05,
-        "16:10" => (ratio - 1.6).abs() < 0.05,
-        "4:3" => (ratio - 1.3333).abs() < 0.05,
-        "1:1" => (ratio - 1.0).abs() < 0.05,
-        _ => true,
-    }
-}
-
-fn selected_resolution(state: &State) -> (u32, u32) {
-    let idx = state
-        .sub_choice_indices_graphics
-        .get(DISPLAY_RESOLUTION_ROW_INDEX)
-        .copied()
-        .unwrap_or(0);
-    state
-        .resolution_choices
-        .get(idx)
-        .copied()
-        .or_else(|| state.resolution_choices.first().copied())
-        .unwrap_or((state.display_width_at_load, state.display_height_at_load))
-}
-
-fn rebuild_refresh_rate_choices(state: &mut State) {
-    if matches!(selected_display_mode(state), DisplayMode::Windowed) {
-        state.refresh_rate_choices = vec![0];
-        if let Some(slot) = state
-            .sub_choice_indices_graphics
-            .get_mut(REFRESH_RATE_ROW_INDEX)
-        {
-            *slot = 0;
-        }
-        if let Some(slot) = state
-            .sub_cursor_indices_graphics
-            .get_mut(REFRESH_RATE_ROW_INDEX)
-        {
-            *slot = 0;
-        }
-        return;
-    }
-
-    let (width, height) = selected_resolution(state);
-    let mon_idx = selected_display_monitor(state);
-    let mut rates = Vec::new();
-
-    // Default choice is always available (0).
-    rates.push(0);
-
-    let supported_rates =
-        display::supported_refresh_rates(state.monitor_specs.get(mon_idx), width, height);
-    rates.extend(supported_rates);
-
-    // Add common fallback rates if list is empty (besides Default)
-    if rates.len() == 1 {
-        rates.extend_from_slice(&[60000, 75000, 120000, 144000, 165000, 240000]);
-    }
-
-    // Preserve current selection if possible, else default to "Default".
-    let current_rate = if let Some(idx) = state
-        .sub_choice_indices_graphics
-        .get(REFRESH_RATE_ROW_INDEX)
-    {
-        state.refresh_rate_choices.get(*idx).copied().unwrap_or(0)
-    } else {
-        0
-    };
-
-    state.refresh_rate_choices = rates;
-
-    let next_idx = state
-        .refresh_rate_choices
-        .iter()
-        .position(|&r| r == current_rate)
-        .unwrap_or(0);
-    if let Some(slot) = state
-        .sub_choice_indices_graphics
-        .get_mut(REFRESH_RATE_ROW_INDEX)
-    {
-        *slot = next_idx;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_graphics
-        .get_mut(REFRESH_RATE_ROW_INDEX)
-    {
-        *slot = next_idx;
-    }
-    if state.max_fps_at_load == 0 && !max_fps_enabled(state) {
-        seed_max_fps_value_choice(state, 0);
-    }
-}
-
-fn rebuild_resolution_choices(state: &mut State, width: u32, height: u32) {
-    let aspect_label = selected_aspect_label(state);
-    let mon_idx = selected_display_monitor(state);
-
-    let mut list: Vec<(u32, u32)> =
-        display::supported_resolutions(state.monitor_specs.get(mon_idx))
-            .into_iter()
-            .filter(|(w, h)| aspect_matches(*w, *h, aspect_label))
-            .collect();
-
-    // 2. If list is empty (e.g. no monitor data or Aspect filter too strict), use presets.
-    if list.is_empty() {
-        list = preset_resolutions_for_aspect(aspect_label);
-    }
-
-    // 3. Keep the current resolution only if it matches the selected aspect.
-    if aspect_matches(width, height, aspect_label) {
-        push_unique_resolution(&mut list, width, height);
-    }
-
-    // Sort descending by width then height (typical UI preference).
-    list.sort_by(|a, b| a.0.cmp(&b.0).then(a.1.cmp(&b.1)));
-
-    state.resolution_choices = list;
-    let next_idx = state
-        .resolution_choices
-        .iter()
-        .position(|&(w, h)| w == width && h == height)
-        .unwrap_or(0);
-    if let Some(slot) = state
-        .sub_choice_indices_graphics
-        .get_mut(DISPLAY_RESOLUTION_ROW_INDEX)
-    {
-        *slot = next_idx;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_graphics
-        .get_mut(DISPLAY_RESOLUTION_ROW_INDEX)
-    {
-        *slot = next_idx;
-    }
-
-    // Rebuild refresh rates since available rates depend on resolution.
-    rebuild_refresh_rate_choices(state);
 }
 
 #[inline(always)]
@@ -5144,8 +1977,6 @@ fn move_submenu_selection_vertical(
     apply_submenu_inline_x_to_row(state, asset_manager, kind, state.sub_selected);
 }
 
-const SOUND_VOLUME_LEVELS: [u8; 6] = [0, 10, 25, 50, 75, 100];
-
 fn set_choice_by_id(choice_indices: &mut Vec<usize>, rows: &[SubRow], id: SubRowId, idx: usize) {
     if let Some(pos) = rows.iter().position(|r| r.id == id)
         && let Some(slot) = choice_indices.get_mut(pos)
@@ -5155,729 +1986,12 @@ fn set_choice_by_id(choice_indices: &mut Vec<usize>, rows: &[SubRow], id: SubRow
     }
 }
 
-fn master_volume_choice_index(volume: u8) -> usize {
-    let mut best_idx = 0usize;
-    let mut best_diff = u8::MAX;
-    for (idx, level) in SOUND_VOLUME_LEVELS.iter().enumerate() {
-        let diff = volume.abs_diff(*level);
-        if diff < best_diff {
-            best_diff = diff;
-            best_idx = idx;
-        }
-    }
-    best_idx
-}
-
-fn master_volume_from_choice(idx: usize) -> u8 {
-    SOUND_VOLUME_LEVELS
-        .get(idx)
-        .copied()
-        .unwrap_or_else(|| *SOUND_VOLUME_LEVELS.last().unwrap_or(&100))
-}
-
-fn sound_row_index(id: SubRowId) -> Option<usize> {
-    SOUND_OPTIONS_ROWS.iter().position(|row| row.id == id)
-}
-
-fn selected_sound_device_choice(state: &State) -> usize {
-    sound_row_index(SubRowId::SoundDevice)
-        .and_then(|idx| state.sub_choice_indices_sound.get(idx).copied())
-        .unwrap_or(0)
-}
-
-fn sound_sample_rate_choices(state: &State) -> Vec<Option<u32>> {
-    let mut choices = Vec::new();
-    choices.push(None);
-    let device_idx =
-        selected_sound_device_choice(state).min(state.sound_device_options.len().saturating_sub(1));
-    if let Some(option) = state.sound_device_options.get(device_idx) {
-        for &hz in &option.sample_rates_hz {
-            let rate = Some(hz);
-            if !choices.contains(&rate) {
-                choices.push(rate);
-            }
-        }
-    }
-    if choices.len() == 1 {
-        choices.push(Some(44100));
-        choices.push(Some(48000));
-    }
-    choices
-}
-
-fn sound_device_choice_index(options: &[SoundDeviceOption], config_index: Option<u16>) -> usize {
-    let Some(target) = config_index else {
-        return 0;
-    };
-    options
-        .iter()
-        .position(|opt| opt.config_index == Some(target))
-        .unwrap_or(0)
-}
-
-fn sound_device_from_choice(state: &State, idx: usize) -> Option<u16> {
-    state
-        .sound_device_options
-        .get(idx)
-        .and_then(|opt| opt.config_index)
-}
-
-fn audio_output_mode_choice_index(mode: config::AudioOutputMode) -> usize {
-    match mode {
-        config::AudioOutputMode::Auto => 0,
-        config::AudioOutputMode::Shared | config::AudioOutputMode::Exclusive => 1,
-    }
-}
-
-fn audio_output_mode_from_choice(idx: usize) -> config::AudioOutputMode {
-    match idx {
-        1 => config::AudioOutputMode::Shared,
-        _ => config::AudioOutputMode::Auto,
-    }
-}
-
-#[cfg(target_os = "linux")]
-#[inline(always)]
-const fn alsa_exclusive_choice_index(mode: config::AudioOutputMode) -> usize {
-    if matches!(mode, config::AudioOutputMode::Exclusive) {
-        1
-    } else {
-        0
-    }
-}
-
-#[cfg(target_os = "linux")]
-#[inline(always)]
-fn selected_audio_output_mode(state: &State) -> config::AudioOutputMode {
-    sound_row_index(SubRowId::AudioOutputMode)
-        .and_then(|idx| state.sub_choice_indices_sound.get(idx).copied())
-        .map(audio_output_mode_from_choice)
-        .unwrap_or(config::AudioOutputMode::Auto)
-}
-
-#[cfg(target_os = "linux")]
-fn linux_audio_backend_choice_index(state: &State, backend: config::LinuxAudioBackend) -> usize {
-    let target = linux_backend_label(backend).to_string();
-    state
-        .linux_backend_choices
-        .iter()
-        .position(|choice| *choice == target)
-        .unwrap_or(0)
-}
-
-#[cfg(target_os = "linux")]
-fn linux_audio_backend_from_choice(state: &State, idx: usize) -> config::LinuxAudioBackend {
-    match state
-        .linux_backend_choices
-        .get(idx)
-        .map(String::as_str)
-        .unwrap_or("Auto")
-    {
-        "PipeWire" => config::LinuxAudioBackend::PipeWire,
-        "PulseAudio" => config::LinuxAudioBackend::PulseAudio,
-        "JACK" => config::LinuxAudioBackend::Jack,
-        "ALSA" => config::LinuxAudioBackend::Alsa,
-        _ => config::LinuxAudioBackend::Auto,
-    }
-}
-
-#[cfg(target_os = "linux")]
-#[inline(always)]
-fn selected_linux_audio_backend(state: &State) -> config::LinuxAudioBackend {
-    sound_row_index(SubRowId::LinuxAudioBackend)
-        .and_then(|idx| state.sub_choice_indices_sound.get(idx).copied())
-        .map(|idx| linux_audio_backend_from_choice(state, idx))
-        .unwrap_or(config::LinuxAudioBackend::Auto)
-}
-
-#[cfg(target_os = "linux")]
-#[inline(always)]
-fn sound_show_alsa_exclusive(state: &State) -> bool {
-    matches!(
-        selected_linux_audio_backend(state),
-        config::LinuxAudioBackend::Alsa
-    )
-}
-
-#[cfg(target_os = "linux")]
-fn sound_parent_row(actual_idx: usize) -> Option<usize> {
-    let child_idx = sound_row_index(SubRowId::AlsaExclusive)?;
-    if actual_idx != child_idx {
-        return None;
-    }
-    sound_row_index(SubRowId::LinuxAudioBackend)
-}
-
-fn set_sound_choice_index(state: &mut State, id: SubRowId, idx: usize) {
-    let Some(row_idx) = sound_row_index(id) else {
-        return;
-    };
-    if let Some(slot) = state.sub_choice_indices_sound.get_mut(row_idx) {
-        *slot = idx;
-    }
-    if let Some(slot) = state.sub_cursor_indices_sound.get_mut(row_idx) {
-        *slot = idx;
-    }
-}
-
-fn sample_rate_choice_index(state: &State, rate: Option<u32>) -> usize {
-    sound_sample_rate_choices(state)
-        .iter()
-        .position(|&r| r == rate)
-        .unwrap_or(0)
-}
-
-fn sample_rate_from_choice(state: &State, idx: usize) -> Option<u32> {
-    sound_sample_rate_choices(state).get(idx).copied().flatten()
-}
-
-fn bg_brightness_choice_index(brightness: f32) -> usize {
-    ((brightness.clamp(0.0, 1.0) * 10.0).round() as i32).clamp(0, 10) as usize
-}
-
-fn bg_brightness_from_choice(idx: usize) -> f32 {
-    idx.min(10) as f32 / 10.0
-}
-
-fn music_wheel_scroll_speed_choice_index(speed: u8) -> usize {
-    let mut best_idx = 0usize;
-    let mut best_diff = u8::MAX;
-    for (idx, value) in MUSIC_WHEEL_SCROLL_SPEED_VALUES.iter().enumerate() {
-        let diff = speed.abs_diff(*value);
-        if diff < best_diff {
-            best_diff = diff;
-            best_idx = idx;
-        }
-    }
-    best_idx
-}
-
-fn music_wheel_scroll_speed_from_choice(idx: usize) -> u8 {
-    MUSIC_WHEEL_SCROLL_SPEED_VALUES
-        .get(idx)
-        .copied()
-        .unwrap_or(15)
-}
-
-#[inline(always)]
-const fn scorebox_cycle_mask(itg: bool, ex: bool, hard_ex: bool, tournaments: bool) -> u8 {
-    (itg as u8) | ((ex as u8) << 1) | ((hard_ex as u8) << 2) | ((tournaments as u8) << 3)
-}
-
-#[inline(always)]
-const fn auto_screenshot_cursor_index(mask: u8) -> usize {
-    if (mask & config::AUTO_SS_PBS) != 0 {
-        0
-    } else if (mask & config::AUTO_SS_FAILS) != 0 {
-        1
-    } else if (mask & config::AUTO_SS_CLEARS) != 0 {
-        2
-    } else if (mask & config::AUTO_SS_QUADS) != 0 {
-        3
-    } else if (mask & config::AUTO_SS_QUINTS) != 0 {
-        4
-    } else {
-        0
-    }
-}
-
-#[inline(always)]
-const fn scorebox_cycle_cursor_index(
-    itg: bool,
-    ex: bool,
-    hard_ex: bool,
-    tournaments: bool,
-) -> usize {
-    if itg {
-        0
-    } else if ex {
-        1
-    } else if hard_ex {
-        2
-    } else if tournaments {
-        3
-    } else {
-        0
-    }
-}
-
-#[inline(always)]
-const fn scorebox_cycle_bit_from_choice(idx: usize) -> u8 {
-    if idx < SELECT_MUSIC_SCOREBOX_CYCLE_NUM_CHOICES {
-        1u8 << (idx as u8)
-    } else {
-        0
-    }
-}
-
-#[inline(always)]
-const fn scorebox_cycle_mask_from_config(cfg: &config::Config) -> u8 {
-    scorebox_cycle_mask(
-        cfg.select_music_scorebox_cycle_itg,
-        cfg.select_music_scorebox_cycle_ex,
-        cfg.select_music_scorebox_cycle_hard_ex,
-        cfg.select_music_scorebox_cycle_tournaments,
-    )
-}
-
-#[inline(always)]
-fn apply_scorebox_cycle_mask(mask: u8) {
-    config::update_select_music_scorebox_cycle_itg((mask & (1u8 << 0)) != 0);
-    config::update_select_music_scorebox_cycle_ex((mask & (1u8 << 1)) != 0);
-    config::update_select_music_scorebox_cycle_hard_ex((mask & (1u8 << 2)) != 0);
-    config::update_select_music_scorebox_cycle_tournaments((mask & (1u8 << 3)) != 0);
-}
-
-fn toggle_select_music_scorebox_cycle_option(state: &mut State, choice_idx: usize) {
-    let bit = scorebox_cycle_bit_from_choice(choice_idx);
-    if bit == 0 {
-        return;
-    }
-    let mut mask = scorebox_cycle_mask_from_config(&config::get());
-    if (mask & bit) != 0 {
-        mask &= !bit;
-    } else {
-        mask |= bit;
-    }
-    apply_scorebox_cycle_mask(mask);
-
-    let clamped = choice_idx.min(SELECT_MUSIC_SCOREBOX_CYCLE_NUM_CHOICES.saturating_sub(1));
-    if let Some(slot) = state
-        .sub_choice_indices_select_music
-        .get_mut(SELECT_MUSIC_SCOREBOX_CYCLE_ROW_INDEX)
-    {
-        *slot = clamped;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_select_music
-        .get_mut(SELECT_MUSIC_SCOREBOX_CYCLE_ROW_INDEX)
-    {
-        *slot = clamped;
-    }
-    audio::play_sfx("assets/sounds/change_value.ogg");
-}
-
-#[inline(always)]
-fn select_music_scorebox_cycle_enabled_mask() -> u8 {
-    scorebox_cycle_mask_from_config(&config::get())
-}
-
-#[inline(always)]
-const fn auto_screenshot_bit_from_choice(idx: usize) -> u8 {
-    config::auto_screenshot_bit(idx)
-}
-
-#[inline(always)]
-fn auto_screenshot_enabled_mask() -> u8 {
-    config::get().auto_screenshot_eval
-}
-
-fn toggle_auto_screenshot_option(state: &mut State, choice_idx: usize) {
-    let bit = auto_screenshot_bit_from_choice(choice_idx);
-    if bit == 0 {
-        return;
-    }
-    let mut mask = config::get().auto_screenshot_eval;
-    if (mask & bit) != 0 {
-        mask &= !bit;
-    } else {
-        mask |= bit;
-    }
-    config::update_auto_screenshot_eval(mask);
-
-    let clamped = choice_idx.min(config::AUTO_SS_NUM_FLAGS.saturating_sub(1));
-    set_choice_by_id(
-        &mut state.sub_choice_indices_gameplay,
-        GAMEPLAY_OPTIONS_ROWS,
-        SubRowId::AutoScreenshot,
-        clamped,
-    );
-    set_choice_by_id(
-        &mut state.sub_cursor_indices_gameplay,
-        GAMEPLAY_OPTIONS_ROWS,
-        SubRowId::AutoScreenshot,
-        clamped,
-    );
-    audio::play_sfx("assets/sounds/change_value.ogg");
-}
-
-const fn breakdown_style_choice_index(style: BreakdownStyle) -> usize {
-    match style {
-        BreakdownStyle::Sl => 0,
-        BreakdownStyle::Sn => 1,
-    }
-}
-
-const fn breakdown_style_from_choice(idx: usize) -> BreakdownStyle {
-    match idx {
-        1 => BreakdownStyle::Sn,
-        _ => BreakdownStyle::Sl,
-    }
-}
-
-const fn default_fail_type_choice_index(fail_type: DefaultFailType) -> usize {
-    match fail_type {
-        DefaultFailType::Immediate => 0,
-        DefaultFailType::ImmediateContinue => 1,
-    }
-}
-
-const fn default_fail_type_from_choice(idx: usize) -> DefaultFailType {
-    match idx {
-        0 => DefaultFailType::Immediate,
-        _ => DefaultFailType::ImmediateContinue,
-    }
-}
-
-const fn sync_graph_mode_choice_index(mode: SyncGraphMode) -> usize {
-    match mode {
-        SyncGraphMode::Frequency => 0,
-        SyncGraphMode::BeatIndex => 1,
-        SyncGraphMode::PostKernelFingerprint => 2,
-    }
-}
-
-const fn sync_graph_mode_from_choice(idx: usize) -> SyncGraphMode {
-    match idx {
-        0 => SyncGraphMode::Frequency,
-        1 => SyncGraphMode::BeatIndex,
-        _ => SyncGraphMode::PostKernelFingerprint,
-    }
-}
-
-const fn sync_confidence_choice_index(percent: u8) -> usize {
-    let capped = if percent > 100 { 100 } else { percent };
-    ((capped as usize) + 2) / 5
-}
-
-const fn sync_confidence_from_choice(idx: usize) -> u8 {
-    let capped = if idx > 20 { 20 } else { idx };
-    capped as u8 * 5
-}
-
-const fn null_or_die_kernel_target_choice_index(target: KernelTarget) -> usize {
-    match target {
-        KernelTarget::Digest => 0,
-        KernelTarget::Accumulator => 1,
-    }
-}
-
-const fn null_or_die_kernel_target_from_choice(idx: usize) -> KernelTarget {
-    match idx {
-        1 => KernelTarget::Accumulator,
-        _ => KernelTarget::Digest,
-    }
-}
-
-const fn null_or_die_kernel_type_choice_index(kind: BiasKernel) -> usize {
-    match kind {
-        BiasKernel::Rising => 0,
-        BiasKernel::Loudest => 1,
-    }
-}
-
-const fn null_or_die_kernel_type_from_choice(idx: usize) -> BiasKernel {
-    match idx {
-        1 => BiasKernel::Loudest,
-        _ => BiasKernel::Rising,
-    }
-}
-
 const fn yes_no_choice_index(enabled: bool) -> usize {
     if enabled { 1 } else { 0 }
 }
 
 const fn yes_no_from_choice(idx: usize) -> bool {
     idx == 1
-}
-
-const fn translated_titles_choice_index(translated_titles: bool) -> usize {
-    if translated_titles { 0 } else { 1 }
-}
-
-const fn translated_titles_from_choice(idx: usize) -> bool {
-    idx == 0
-}
-
-const fn language_choice_index(flag: config::LanguageFlag) -> usize {
-    match flag {
-        config::LanguageFlag::Auto | config::LanguageFlag::English => 0,
-        config::LanguageFlag::German => 1,
-        config::LanguageFlag::Spanish => 2,
-        config::LanguageFlag::French => 3,
-        config::LanguageFlag::Italian => 4,
-        config::LanguageFlag::Japanese => 5,
-        config::LanguageFlag::Polish => 6,
-        config::LanguageFlag::PortugueseBrazil => 7,
-        config::LanguageFlag::Russian => 8,
-        config::LanguageFlag::Swedish => 9,
-        config::LanguageFlag::Pseudo => 10,
-    }
-}
-
-const fn language_flag_from_choice(idx: usize) -> config::LanguageFlag {
-    match idx {
-        1 => config::LanguageFlag::German,
-        2 => config::LanguageFlag::Spanish,
-        3 => config::LanguageFlag::French,
-        4 => config::LanguageFlag::Italian,
-        5 => config::LanguageFlag::Japanese,
-        6 => config::LanguageFlag::Polish,
-        7 => config::LanguageFlag::PortugueseBrazil,
-        8 => config::LanguageFlag::Russian,
-        9 => config::LanguageFlag::Swedish,
-        10 => config::LanguageFlag::Pseudo,
-        _ => config::LanguageFlag::English,
-    }
-}
-
-const fn select_music_pattern_info_choice_index(mode: SelectMusicPatternInfoMode) -> usize {
-    match mode {
-        SelectMusicPatternInfoMode::Auto => 0,
-        SelectMusicPatternInfoMode::Tech => 1,
-        SelectMusicPatternInfoMode::Stamina => 2,
-    }
-}
-
-const fn select_music_pattern_info_from_choice(idx: usize) -> SelectMusicPatternInfoMode {
-    match idx {
-        1 => SelectMusicPatternInfoMode::Tech,
-        2 => SelectMusicPatternInfoMode::Stamina,
-        _ => SelectMusicPatternInfoMode::Auto,
-    }
-}
-
-#[inline(always)]
-const fn select_music_chart_info_mask(peak_nps: bool, matrix_rating: bool) -> u8 {
-    (peak_nps as u8) | ((matrix_rating as u8) << 1)
-}
-
-#[inline(always)]
-const fn select_music_chart_info_cursor_index(peak_nps: bool, matrix_rating: bool) -> usize {
-    if peak_nps {
-        0
-    } else if matrix_rating {
-        1
-    } else {
-        0
-    }
-}
-
-#[inline(always)]
-const fn select_music_chart_info_bit_from_choice(idx: usize) -> u8 {
-    if idx < SELECT_MUSIC_CHART_INFO_NUM_CHOICES {
-        1u8 << (idx as u8)
-    } else {
-        0
-    }
-}
-
-#[inline(always)]
-const fn select_music_chart_info_mask_from_config(cfg: &config::Config) -> u8 {
-    select_music_chart_info_mask(
-        cfg.select_music_chart_info_peak_nps,
-        cfg.select_music_chart_info_matrix_rating,
-    )
-}
-
-#[inline(always)]
-fn apply_select_music_chart_info_mask(mask: u8) {
-    config::update_select_music_chart_info_peak_nps((mask & (1u8 << 0)) != 0);
-    config::update_select_music_chart_info_matrix_rating((mask & (1u8 << 1)) != 0);
-}
-
-fn toggle_select_music_chart_info_option(state: &mut State, choice_idx: usize) {
-    let bit = select_music_chart_info_bit_from_choice(choice_idx);
-    if bit == 0 {
-        return;
-    }
-    let mut mask = select_music_chart_info_mask_from_config(&config::get());
-    if (mask & bit) != 0 {
-        if (mask & !bit) == 0 {
-            return;
-        }
-        mask &= !bit;
-    } else {
-        mask |= bit;
-    }
-    apply_select_music_chart_info_mask(mask);
-
-    let clamped = choice_idx.min(SELECT_MUSIC_CHART_INFO_NUM_CHOICES.saturating_sub(1));
-    if let Some(slot) = state
-        .sub_choice_indices_select_music
-        .get_mut(SELECT_MUSIC_CHART_INFO_ROW_INDEX)
-    {
-        *slot = clamped;
-    }
-    if let Some(slot) = state
-        .sub_cursor_indices_select_music
-        .get_mut(SELECT_MUSIC_CHART_INFO_ROW_INDEX)
-    {
-        *slot = clamped;
-    }
-    audio::play_sfx("assets/sounds/change_value.ogg");
-}
-
-#[inline(always)]
-fn select_music_chart_info_enabled_mask() -> u8 {
-    let mask = select_music_chart_info_mask_from_config(&config::get());
-    if mask == 0 { 1 } else { mask }
-}
-
-const fn select_music_itl_wheel_choice_index(mode: SelectMusicItlWheelMode) -> usize {
-    match mode {
-        SelectMusicItlWheelMode::Off => 0,
-        SelectMusicItlWheelMode::Score => 1,
-        SelectMusicItlWheelMode::PointsAndScore => 2,
-    }
-}
-
-const fn select_music_itl_rank_choice_index(mode: SelectMusicItlRankMode) -> usize {
-    match mode {
-        SelectMusicItlRankMode::None => 0,
-        SelectMusicItlRankMode::Chart => 1,
-        SelectMusicItlRankMode::Overall => 2,
-    }
-}
-
-const fn select_music_itl_rank_from_choice(idx: usize) -> SelectMusicItlRankMode {
-    match idx {
-        1 => SelectMusicItlRankMode::Chart,
-        2 => SelectMusicItlRankMode::Overall,
-        _ => SelectMusicItlRankMode::None,
-    }
-}
-
-const fn select_music_itl_wheel_from_choice(idx: usize) -> SelectMusicItlWheelMode {
-    match idx {
-        1 => SelectMusicItlWheelMode::Score,
-        2 => SelectMusicItlWheelMode::PointsAndScore,
-        _ => SelectMusicItlWheelMode::Off,
-    }
-}
-
-const fn select_music_wheel_style_choice_index(style: SelectMusicWheelStyle) -> usize {
-    match style {
-        SelectMusicWheelStyle::Itg => 0,
-        SelectMusicWheelStyle::Iidx => 1,
-    }
-}
-
-const fn select_music_wheel_style_from_choice(idx: usize) -> SelectMusicWheelStyle {
-    match idx {
-        1 => SelectMusicWheelStyle::Iidx,
-        _ => SelectMusicWheelStyle::Itg,
-    }
-}
-
-const fn new_pack_mode_choice_index(mode: NewPackMode) -> usize {
-    match mode {
-        NewPackMode::Disabled => 0,
-        NewPackMode::OpenPack => 1,
-        NewPackMode::HasScore => 2,
-    }
-}
-
-const fn new_pack_mode_from_choice(idx: usize) -> NewPackMode {
-    match idx {
-        1 => NewPackMode::OpenPack,
-        2 => NewPackMode::HasScore,
-        _ => NewPackMode::Disabled,
-    }
-}
-
-const fn select_music_scorebox_placement_choice_index(
-    placement: SelectMusicScoreboxPlacement,
-) -> usize {
-    match placement {
-        SelectMusicScoreboxPlacement::Auto => 0,
-        SelectMusicScoreboxPlacement::StepPane => 1,
-    }
-}
-
-const fn select_music_scorebox_placement_from_choice(idx: usize) -> SelectMusicScoreboxPlacement {
-    match idx {
-        1 => SelectMusicScoreboxPlacement::StepPane,
-        _ => SelectMusicScoreboxPlacement::Auto,
-    }
-}
-
-const fn machine_preferred_style_choice_index(style: MachinePreferredPlayStyle) -> usize {
-    match style {
-        MachinePreferredPlayStyle::Single => 0,
-        MachinePreferredPlayStyle::Versus => 1,
-        MachinePreferredPlayStyle::Double => 2,
-    }
-}
-
-const fn machine_preferred_style_from_choice(idx: usize) -> MachinePreferredPlayStyle {
-    match idx {
-        1 => MachinePreferredPlayStyle::Versus,
-        2 => MachinePreferredPlayStyle::Double,
-        _ => MachinePreferredPlayStyle::Single,
-    }
-}
-
-const fn machine_preferred_mode_choice_index(mode: MachinePreferredPlayMode) -> usize {
-    match mode {
-        MachinePreferredPlayMode::Regular => 0,
-        MachinePreferredPlayMode::Marathon => 1,
-    }
-}
-
-const fn machine_preferred_mode_from_choice(idx: usize) -> MachinePreferredPlayMode {
-    match idx {
-        1 => MachinePreferredPlayMode::Marathon,
-        _ => MachinePreferredPlayMode::Regular,
-    }
-}
-
-const fn machine_font_choice_index(font: MachineFont) -> usize {
-    match font {
-        MachineFont::Common => 0,
-        MachineFont::Mega => 1,
-    }
-}
-
-const fn machine_font_from_choice(idx: usize) -> MachineFont {
-    match idx {
-        1 => MachineFont::Mega,
-        _ => MachineFont::Common,
-    }
-}
-
-const fn menu_background_style_choice_index(style: MenuBackgroundStyle) -> usize {
-    match style {
-        MenuBackgroundStyle::Hearts => 0,
-        MenuBackgroundStyle::Technique => 1,
-    }
-}
-
-const fn menu_background_style_from_choice(idx: usize) -> MenuBackgroundStyle {
-    match idx {
-        1 => MenuBackgroundStyle::Technique,
-        _ => MenuBackgroundStyle::Hearts,
-    }
-}
-
-const fn log_level_choice_index(level: LogLevel) -> usize {
-    match level {
-        LogLevel::Error => 0,
-        LogLevel::Warn => 1,
-        LogLevel::Info => 2,
-        LogLevel::Debug => 3,
-        LogLevel::Trace => 4,
-    }
-}
-
-const fn log_level_from_choice(idx: usize) -> LogLevel {
-    match idx {
-        0 => LogLevel::Error,
-        1 => LogLevel::Warn,
-        2 => LogLevel::Info,
-        3 => LogLevel::Debug,
-        _ => LogLevel::Trace,
-    }
 }
 
 pub struct State {
