+++ src/screens/options/submenus/mod.rs
@@ -1,0 +1,43 @@
+use super::*;
+
+mod system;
+mod graphics;
+mod input_dev;
+mod machine;
+mod course;
+mod gameplay;
+mod sound;
+mod select_music;
+mod advanced;
+mod online;
+mod pack_sync;
+mod score_import;
+mod null_or_die;
+
+#[allow(unused_imports)]
+pub(super) use system::*;
+#[allow(unused_imports)]
+pub(super) use graphics::*;
+pub use graphics::update_monitor_specs;
+#[allow(unused_imports)]
+pub(super) use input_dev::*;
+#[allow(unused_imports)]
+pub(super) use machine::*;
+#[allow(unused_imports)]
+pub(super) use course::*;
+#[allow(unused_imports)]
+pub(super) use gameplay::*;
+#[allow(unused_imports)]
+pub(super) use sound::*;
+#[allow(unused_imports)]
+pub(super) use select_music::*;
+#[allow(unused_imports)]
+pub(super) use advanced::*;
+#[allow(unused_imports)]
+pub(super) use online::*;
+#[allow(unused_imports)]
+pub(super) use pack_sync::*;
+#[allow(unused_imports)]
+pub(super) use score_import::*;
+#[allow(unused_imports)]
+pub(super) use null_or_die::*;
